% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_MicroEcoTools_df.R
\name{convert_to_MicroEcoTools_df}
\alias{convert_to_MicroEcoTools_df}
\title{Convert SummarizedExperiment or phyloseq Object to a MicroEcoTools-Compatible Data Frame}
\usage{
convert_to_MicroEcoTools_df(
  x,
  group_col = "Exp.Grp",
  replicate_col = "Replicate",
  assay_name = "counts"
)
}
\arguments{
\item{x}{A SummarizedExperiment or phyloseq object.}

\item{group_col}{A character string indicating the column name in the sample metadata that
represents the experimental group. Default is `"Exp.Grp"`.}

\item{replicate_col}{A character string indicating the column name in the sample metadata that
represents replicates. Default is `"Replicate"`.}

\item{assay_name}{For SummarizedExperiment objects, the name of the assay to use. Default is `"counts"`.}
}
\value{
A data frame with the first two columns corresponding to the experimental group and replicate,
  and the remaining columns containing the taxa/feature counts.
}
\description{
This function takes in either a [SummarizedExperiment::SummarizedExperiment()] or a
[phyloseq::phyloseq()] object and converts it to a data frame formatted for MicroEcoTools analyses.
The returned data frame will have two metadata columns (by default, `Exp.Grp` and `Replicate`)
followed by the taxa/feature count data.
}
\examples{
\dontrun{
  ## For a SummarizedExperiment object:
  library(SummarizedExperiment)
  # Suppose 'se_obj' is a SummarizedExperiment with colData containing columns "Exp.Grp" and "Replicate"
  df <- convert_to_microecotools_df(se_obj, group_col = "Exp.Grp", replicate_col = "Replicate")

  ## For a phyloseq object:
  library(phyloseq)
  # Suppose 'ps_obj' is a phyloseq object with sample_data containing columns "Exp.Grp" and "Replicate"
  df <- convert_to_microecotools_df(ps_obj, group_col = "Exp.Grp", replicate_col = "Replicate")
}

}
