% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{CSR_IP2G_data}
\alias{CSR_IP2G_data}
\title{CSR_IP2G_data}
\format{
A data frame with 24 rows and 59 variables:
\describe{
  \item{Exp.Grp}{Experimental group identifier (integer).}
  \item{Replicate}{Replicate number for each experimental group (integer).}
  \item{biosynthetic.process}{Abundance of the biosynthetic process (numeric).}
  \item{carbohydrate.metabolic.process}{Abundance of the carbohydrate metabolic process (numeric).}
  \item{catabolic.process}{Abundance of the catabolic process (numeric).}
  \item{cell.communication}{Abundance of the cell communication process (numeric).}
  \item{cell.division}{Abundance of the cell division process (numeric).}
  \item{cell.motility}{Abundance of the cell motility process (numeric).}
  \item{cellular.amino.acid.metabolic.process}{Abundance of the cellular amino acid metabolic process (numeric).}
  \item{cellular.component.organization}{Abundance of the cellular component organization process (numeric).}
  \item{cellular.respiration}{Abundance of the cellular respiration process (numeric).}
  \item{chemotaxis}{Abundance of the chemotaxis process (numeric).}
  \item{cofactor.metabolic.process}{Abundance of the cofactor metabolic process (numeric).}
  \item{conjugation}{Abundance of the conjugation process (numeric).}
  \item{DNA.metabolic.process}{Abundance of the DNA metabolic process (numeric).}
  \item{electron.transport.chain}{Abundance of the electron transport chain (numeric).}
  \item{generation.of.precursor.metabolites.and.energy}{Abundance of the generation of precursor metabolites and energy process (numeric).}
  \item{homeostatic.process}{Abundance of the homeostatic process (numeric).}
  \item{lipid.metabolic.process}{Abundance of the lipid metabolic process (numeric).}
  \item{nitrogen.compound.metabolic.process}{Abundance of the nitrogen compound metabolic process (numeric).}
  \item{nucleotide.metabolic.process}{Abundance of the nucleotide metabolic process (numeric).}
  \item{oxidation.reduction.process}{Abundance of the oxidation-reduction process (numeric).}
  \item{phosphorylation}{Abundance of the phosphorylation process (numeric).}
  \item{protein.metabolic.process}{Abundance of the protein metabolic process (numeric).}
  \item{response.to.stress}{Abundance of the response to stress (numeric).}
  \item{RNA.metabolic.process}{Abundance of the RNA metabolic process (numeric).}
  \item{signal.transduction}{Abundance of the signal transduction process (numeric).}
  \item{small.molecule.metabolic.process}{Abundance of the small molecule metabolic process (numeric).}
  \item{sporulation}{Abundance of the sporulation process (numeric).}
  \item{sulfur.compound.metabolic.process}{Abundance of the sulfur compound metabolic process (numeric).}
  \item{transcription..DNA.templated}{Abundance of the transcription, DNA-templated process (numeric).}
  \item{translation}{Abundance of the translation process (numeric).}
  \item{transport}{Abundance of the transport process (numeric).}
  \item{transposition}{Abundance of the transposition process (numeric).}
  \item{cofactor.binding}{Abundance of the cofactor binding activity (numeric).}
  \item{enzyme.regulator.activity}{Abundance of the enzyme regulator activity (numeric).}
  \item{hydrolase.activity}{Abundance of the hydrolase activity (numeric).}
  \item{isomerase.activity}{Abundance of the isomerase activity (numeric).}
  \item{ligase.activity}{Abundance of the ligase activity (numeric).}
  \item{lyase.activity}{Abundance of the lyase activity (numeric).}
  \item{metal.ion.binding}{Abundance of the metal ion binding activity (numeric).}
  \item{nucleic.acid.binding}{Abundance of the nucleic acid binding activity (numeric).}
  \item{nucleotide.binding}{Abundance of the nucleotide binding activity (numeric).}
  \item{oxidoreductase.activity}{Abundance of the oxidoreductase activity (numeric).}
  \item{phosphorelay.response.regulator.activity}{Abundance of the phosphorelay response regulator activity (numeric).}
  \item{receptor.activity}{Abundance of the receptor activity (numeric).}
  \item{transcription.factor.activity..protein.binding}{Abundance of the transcription factor activity with protein binding (numeric).}
  \item{transferase.activity}{Abundance of the transferase activity (numeric).}
  \item{transporter.activity}{Abundance of the transporter activity (numeric).}
  \item{unfolded.protein.binding}{Abundance of the unfolded protein binding activity (numeric).}
  \item{ATP.binding.cassette..ABC..transporter.complex}{Abundance of the ATP-binding cassette (ABC) transporter complex (numeric).}
  \item{cell.wall}{Abundance of the cell wall (numeric).}
  \item{chromosome}{Abundance of the chromosome (numeric).}
  \item{cytoplasm}{Abundance of the cytoplasm (numeric).}
  \item{DNA.polymerase.complex}{Abundance of the DNA polymerase complex (numeric).}
  \item{membrane}{Abundance of the membrane (numeric).}
  \item{periplasmic.space}{Abundance of the periplasmic space (numeric).}
  \item{ribosome}{Abundance of the ribosome (numeric).}
}
}
\source{
Experimental data from CSR IP2G metagenomics study.
}
\usage{
CSR_IP2G_data
}
\description{
A dataset containing metagenomic trait abundances from a 3CA microcosm experiment. This dataset includes 24 observations and 59 variables related to different biological processes and functions.
}
\keyword{datasets}
