% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{CSR_plot}
\alias{CSR_plot}
\title{Generate a CSR Plot from Plotting Parameters in the result object generated by the CSR_assign}
\usage{
CSR_plot(CSR_results, plot_type = c("abundance", "p_value"))
}
\arguments{
\item{CSR_results}{A list containing two sub-lists:
\itemize{
  \item \code{CSR_plot_sorted_abundance}: A list of plotting parameters for a sorted abundance plot.
  \item \code{CSR_plot_sorted_p_value}: A list of plotting parameters for a sorted adjusted P-value plot.
}
Each sub-list must include:
\itemize{
  \item \code{DaTa}: A data frame containing the data (including metadata and counts).
  \item \code{CSR_cat}: A data frame containing CSR assignment results.
  \item \code{CSR_vis_list}: A vector of column names (character) to be visualized.
  \item \code{var.name}: A character string specifying the variable name (e.g., "Taxa" or "Trait").
  \item \code{sort.var}: A character string specifying the sorting criterion (e.g., "abundance" or "adjusted P-value").
}}

\item{plot_type}{A character string to determine which plot to generate.
Valid options are \code{"abundance"} (default) for the sorted abundance plot, or \code{"p_value"}
for the sorted adjusted P-value plot.}
}
\value{
A ggplot object displaying the CSR plot.
}
\description{
This function generates a ggplot object for visualizing CSR (Competitor–Stress tolerant–Ruderal)
assignments. It accepts the csr_result object generated by the CSR_assign function and a character argument \code{plot_type}
to determine whether to use the parameters for a sorted abundance plot or a sorted adjusted P-value plot.
}
\examples{
\dontrun{

  # Generate and display the sorted abundance plot:
  CSR_plot(csr_assignment, plot_type = "abundance")

  # Generate and display the sorted adjusted P-value plot:
  CSR_plot(csr_assignment, plot_type = "p_value")
}
}
