% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{CSR_TAXA_data}
\alias{CSR_TAXA_data}
\title{CSR_TAXA_data Dataset}
\format{
A data frame with 24 rows and 1509 variables:
\describe{
  \item{Exp.Grp (integer)}{Abundance of this taxa/trait}
  \item{Replicate (integer)}{Abundance of this taxa/trait}
  \item{Nakamurella (numeric)}{Abundance of this taxa/trait}
  \item{Microlunatus (numeric)}{Abundance of this taxa/trait}
  \item{Tetrasphaera (numeric)}{Abundance of this taxa/trait}
  \item{Kouleothrix (numeric)}{Abundance of this taxa/trait}
  \item{Leucobacter (numeric)}{Abundance of this taxa/trait}
  \item{Mycobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Gemmatimonas (numeric)}{Abundance of this taxa/trait}
  \item{Thauera (numeric)}{Abundance of this taxa/trait}
  \item{Caldilinea (numeric)}{Abundance of this taxa/trait}
  \item{Arthrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Microbacterium (numeric)}{Abundance of this taxa/trait}
  \item{Labilithrix (numeric)}{Abundance of this taxa/trait}
  \item{Kineosphaera (numeric)}{Abundance of this taxa/trait}
  \item{Nitrospira (numeric)}{Abundance of this taxa/trait}
  \item{Paracoccus (numeric)}{Abundance of this taxa/trait}
  \item{Dehalobacter (numeric)}{Abundance of this taxa/trait}
  \item{Mesorhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Niabella (numeric)}{Abundance of this taxa/trait}
  \item{Bosea (numeric)}{Abundance of this taxa/trait}
  \item{bacterium.SIT5 (numeric)}{Abundance of this taxa/trait}
  \item{Nitrosomonas (numeric)}{Abundance of this taxa/trait}
  \item{Pyrinomonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Contendobacter (numeric)}{Abundance of this taxa/trait}
  \item{Gemmata (numeric)}{Abundance of this taxa/trait}
  \item{Alicycliphilus (numeric)}{Abundance of this taxa/trait}
  \item{Nocardioides (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Actinobacteria..class...miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Bradyrhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Propionicicella (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Bacteria. (numeric)}{Abundance of this taxa/trait}
  \item{Tessaracoccus (numeric)}{Abundance of this taxa/trait}
  \item{Pseudomonas (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Gammaproteobacteria..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Accumulibacter (numeric)}{Abundance of this taxa/trait}
  \item{Flavobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Roseiflexus (numeric)}{Abundance of this taxa/trait}
  \item{Streptomyces (numeric)}{Abundance of this taxa/trait}
  \item{Pedobacter (numeric)}{Abundance of this taxa/trait}
  \item{Sediminibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Acidovorax (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Competibacter (numeric)}{Abundance of this taxa/trait}
  \item{Propionibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Shinella (numeric)}{Abundance of this taxa/trait}
  \item{Stenotrophomonas (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Betaproteobacteria..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Hassallia (numeric)}{Abundance of this taxa/trait}
  \item{Cellulomonas (numeric)}{Abundance of this taxa/trait}
  \item{Hyphomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Comamonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Saccharimonas (numeric)}{Abundance of this taxa/trait}
  \item{Flavihumibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Saccharibacteria (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Anaerolineae (numeric)}{Abundance of this taxa/trait}
  \item{Singulisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Chryseobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Devosia (numeric)}{Abundance of this taxa/trait}
  \item{Bdellovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Pedosphaera (numeric)}{Abundance of this taxa/trait}
  \item{Terrimonas (numeric)}{Abundance of this taxa/trait}
  \item{Achromobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ottowia (numeric)}{Abundance of this taxa/trait}
  \item{Fluviicola (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Solibacter (numeric)}{Abundance of this taxa/trait}
  \item{Rhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Ideonella (numeric)}{Abundance of this taxa/trait}
  \item{Geminicoccus (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Chitinophagaceae (numeric)}{Abundance of this taxa/trait}
  \item{Haliscomenobacter (numeric)}{Abundance of this taxa/trait}
  \item{Verrucomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Runella (numeric)}{Abundance of this taxa/trait}
  \item{Methylibium (numeric)}{Abundance of this taxa/trait}
  \item{Ardenticatena (numeric)}{Abundance of this taxa/trait}
  \item{Rhodococcus (numeric)}{Abundance of this taxa/trait}
  \item{Fimbriimonas (numeric)}{Abundance of this taxa/trait}
  \item{Prosthecomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Pseudopropionibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Pseudoxanthomonas (numeric)}{Abundance of this taxa/trait}
  \item{Methylobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Zavarzinella (numeric)}{Abundance of this taxa/trait}
  \item{Rubrivivax (numeric)}{Abundance of this taxa/trait}
  \item{Gordonia (numeric)}{Abundance of this taxa/trait}
  \item{Brevundimonas (numeric)}{Abundance of this taxa/trait}
  \item{Sphaerobacter (numeric)}{Abundance of this taxa/trait}
  \item{Segetibacter (numeric)}{Abundance of this taxa/trait}
  \item{Turneriella (numeric)}{Abundance of this taxa/trait}
  \item{Granulicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Nitrolancea (numeric)}{Abundance of this taxa/trait}
  \item{Bryobacter (numeric)}{Abundance of this taxa/trait}
  \item{Geothrix (numeric)}{Abundance of this taxa/trait}
  \item{Actinomyces (numeric)}{Abundance of this taxa/trait}
  \item{Cupriavidus (numeric)}{Abundance of this taxa/trait}
  \item{bacterium.JKG1 (numeric)}{Abundance of this taxa/trait}
  \item{Gemmatirosa (numeric)}{Abundance of this taxa/trait}
  \item{Sphingobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Thermomonas (numeric)}{Abundance of this taxa/trait}
  \item{Agromyces (numeric)}{Abundance of this taxa/trait}
  \item{Azoarcus (numeric)}{Abundance of this taxa/trait}
  \item{Phaeodactylibacter (numeric)}{Abundance of this taxa/trait}
  \item{Lewinella (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Acidobacteriaceae (numeric)}{Abundance of this taxa/trait}
  \item{Burkholderia (numeric)}{Abundance of this taxa/trait}
  \item{Reyranella (numeric)}{Abundance of this taxa/trait}
  \item{Bacteroides (numeric)}{Abundance of this taxa/trait}
  \item{Variovorax (numeric)}{Abundance of this taxa/trait}
  \item{Niastella (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Burkholderiales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Rudaea (numeric)}{Abundance of this taxa/trait}
  \item{Sphingomonas (numeric)}{Abundance of this taxa/trait}
  \item{Schlesneria (numeric)}{Abundance of this taxa/trait}
  \item{Dechloromonas (numeric)}{Abundance of this taxa/trait}
  \item{Intrasporangium (numeric)}{Abundance of this taxa/trait}
  \item{Sorangium (numeric)}{Abundance of this taxa/trait}
  \item{Herpetosiphon (numeric)}{Abundance of this taxa/trait}
  \item{Leptospira (numeric)}{Abundance of this taxa/trait}
  \item{Dyadobacter (numeric)}{Abundance of this taxa/trait}
  \item{Mizugakiibacter (numeric)}{Abundance of this taxa/trait}
  \item{Sulfuritalea (numeric)}{Abundance of this taxa/trait}
  \item{Anaerolinea (numeric)}{Abundance of this taxa/trait}
  \item{Afipia (numeric)}{Abundance of this taxa/trait}
  \item{Corynebacterium (numeric)}{Abundance of this taxa/trait}
  \item{Chloroflexus (numeric)}{Abundance of this taxa/trait}
  \item{Oscillochloris (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Microthrix (numeric)}{Abundance of this taxa/trait}
  \item{Lysobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ilumatobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ktedonobacter (numeric)}{Abundance of this taxa/trait}
  \item{Sphingopyxis (numeric)}{Abundance of this taxa/trait}
  \item{Pseudonocardia (numeric)}{Abundance of this taxa/trait}
  \item{Chloracidobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Mobilicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Rhodopirellula (numeric)}{Abundance of this taxa/trait}
  \item{Pirellula (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Propionibacteriaceae (numeric)}{Abundance of this taxa/trait}
  \item{Rhizobacter (numeric)}{Abundance of this taxa/trait}
  \item{Phycicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Anaeromyxobacter (numeric)}{Abundance of this taxa/trait}
  \item{Opitutus (numeric)}{Abundance of this taxa/trait}
  \item{Leadbetterella (numeric)}{Abundance of this taxa/trait}
  \item{Chondromyces (numeric)}{Abundance of this taxa/trait}
  \item{Brachybacterium (numeric)}{Abundance of this taxa/trait}
  \item{Elioraea (numeric)}{Abundance of this taxa/trait}
  \item{Chlamydia (numeric)}{Abundance of this taxa/trait}
  \item{Microvirga (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Entotheonella (numeric)}{Abundance of this taxa/trait}
  \item{Lentimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Methylosarcina (numeric)}{Abundance of this taxa/trait}
  \item{Frankia (numeric)}{Abundance of this taxa/trait}
  \item{Arsenicicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Deinococcus (numeric)}{Abundance of this taxa/trait}
  \item{Novosphingobium (numeric)}{Abundance of this taxa/trait}
  \item{Nocardia (numeric)}{Abundance of this taxa/trait}
  \item{Paenibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Sandaracinus (numeric)}{Abundance of this taxa/trait}
  \item{Kocuria (numeric)}{Abundance of this taxa/trait}
  \item{Fulvivirga (numeric)}{Abundance of this taxa/trait}
  \item{bacterium.YEK0313 (numeric)}{Abundance of this taxa/trait}
  \item{Knoellia (numeric)}{Abundance of this taxa/trait}
  \item{Chthoniobacter (numeric)}{Abundance of this taxa/trait}
  \item{Xanthomonas (numeric)}{Abundance of this taxa/trait}
  \item{Rhodanobacter (numeric)}{Abundance of this taxa/trait}
  \item{Acidipropionibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Aestuariimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Capnocytophaga (numeric)}{Abundance of this taxa/trait}
  \item{Brevibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Bordetella (numeric)}{Abundance of this taxa/trait}
  \item{Janthinobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Gimesia (numeric)}{Abundance of this taxa/trait}
  \item{Longilinea (numeric)}{Abundance of this taxa/trait}
  \item{Solitalea (numeric)}{Abundance of this taxa/trait}
  \item{Agrobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Amycolatopsis (numeric)}{Abundance of this taxa/trait}
  \item{Rhodopseudomonas (numeric)}{Abundance of this taxa/trait}
  \item{Azospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Spirosoma (numeric)}{Abundance of this taxa/trait}
  \item{Actinoplanes (numeric)}{Abundance of this taxa/trait}
  \item{Polaromonas (numeric)}{Abundance of this taxa/trait}
  \item{Rhodobacter (numeric)}{Abundance of this taxa/trait}
  \item{Sphingobium (numeric)}{Abundance of this taxa/trait}
  \item{Kaistia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Flavobacteriales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Chitinophaga (numeric)}{Abundance of this taxa/trait}
  \item{Meiothermus (numeric)}{Abundance of this taxa/trait}
  \item{Thermobaculum (numeric)}{Abundance of this taxa/trait}
  \item{Ignavibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Pseudolabrys (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Symbiothrix (numeric)}{Abundance of this taxa/trait}
  \item{Elizabethkingia (numeric)}{Abundance of this taxa/trait}
  \item{Myroides (numeric)}{Abundance of this taxa/trait}
  \item{Blastopirellula (numeric)}{Abundance of this taxa/trait}
  \item{Thiobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Geobacter (numeric)}{Abundance of this taxa/trait}
  \item{Aeromicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Curvibacter (numeric)}{Abundance of this taxa/trait}
  \item{Hoeflea (numeric)}{Abundance of this taxa/trait}
  \item{Bacillus (numeric)}{Abundance of this taxa/trait}
  \item{Marmoricola (numeric)}{Abundance of this taxa/trait}
  \item{Algoriphagus (numeric)}{Abundance of this taxa/trait}
  \item{Leifsonia (numeric)}{Abundance of this taxa/trait}
  \item{Aureispira (numeric)}{Abundance of this taxa/trait}
  \item{Fermentimonas (numeric)}{Abundance of this taxa/trait}
  \item{Arenimonas (numeric)}{Abundance of this taxa/trait}
  \item{Sinorhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Geofilum (numeric)}{Abundance of this taxa/trait}
  \item{Pelomonas (numeric)}{Abundance of this taxa/trait}
  \item{Legionella (numeric)}{Abundance of this taxa/trait}
  \item{Thioalkalivibrio (numeric)}{Abundance of this taxa/trait}
  \item{Leptothrix (numeric)}{Abundance of this taxa/trait}
  \item{Clostridium (numeric)}{Abundance of this taxa/trait}
  \item{Rhodovulum (numeric)}{Abundance of this taxa/trait}
  \item{Ochrobactrum (numeric)}{Abundance of this taxa/trait}
  \item{Desulfovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Sporocytophaga (numeric)}{Abundance of this taxa/trait}
  \item{Levilinea (numeric)}{Abundance of this taxa/trait}
  \item{Hydrogenophaga (numeric)}{Abundance of this taxa/trait}
  \item{Pseudorhodoferax (numeric)}{Abundance of this taxa/trait}
  \item{Hymenobacter (numeric)}{Abundance of this taxa/trait}
  \item{Methyloversatilis (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Phycisphaerae (numeric)}{Abundance of this taxa/trait}
  \item{Xanthobacter (numeric)}{Abundance of this taxa/trait}
  \item{Methylocystis (numeric)}{Abundance of this taxa/trait}
  \item{Haliangium (numeric)}{Abundance of this taxa/trait}
  \item{Roseomonas (numeric)}{Abundance of this taxa/trait}
  \item{Belnapia (numeric)}{Abundance of this taxa/trait}
  \item{Pontibacter (numeric)}{Abundance of this taxa/trait}
  \item{Isosphaera (numeric)}{Abundance of this taxa/trait}
  \item{Serinicoccus (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Chloroflexi..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Dongia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhodospirillales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Streptococcus (numeric)}{Abundance of this taxa/trait}
  \item{Aquincola (numeric)}{Abundance of this taxa/trait}
  \item{Aphanocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Myxococcus (numeric)}{Abundance of this taxa/trait}
  \item{Holophaga (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Sphingobacteriales (numeric)}{Abundance of this taxa/trait}
  \item{Ralstonia (numeric)}{Abundance of this taxa/trait}
  \item{Massilia (numeric)}{Abundance of this taxa/trait}
  \item{Aquimarina (numeric)}{Abundance of this taxa/trait}
  \item{Bifidobacterium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Parcubacteria.group (numeric)}{Abundance of this taxa/trait}
  \item{Gemmobacter (numeric)}{Abundance of this taxa/trait}
  \item{Micrococcus (numeric)}{Abundance of this taxa/trait}
  \item{Emticicia (numeric)}{Abundance of this taxa/trait}
  \item{Skermanella (numeric)}{Abundance of this taxa/trait}
  \item{Phenylobacterium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Dehalococcoidia (numeric)}{Abundance of this taxa/trait}
  \item{Beutenbergia (numeric)}{Abundance of this taxa/trait}
  \item{Jiangella (numeric)}{Abundance of this taxa/trait}
  \item{Pimelobacter (numeric)}{Abundance of this taxa/trait}
  \item{Luteimonas (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Comamonadaceae (numeric)}{Abundance of this taxa/trait}
  \item{Nitratireductor (numeric)}{Abundance of this taxa/trait}
  \item{Exiguobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Conexibacter (numeric)}{Abundance of this taxa/trait}
  \item{Nitrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Actinotalea (numeric)}{Abundance of this taxa/trait}
  \item{Luteipulveratus (numeric)}{Abundance of this taxa/trait}
  \item{Kribbella (numeric)}{Abundance of this taxa/trait}
  \item{Inquilinus (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Alphaproteobacteria..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Paraburkholderia (numeric)}{Abundance of this taxa/trait}
  \item{Solirubrobacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Verrucomicrobia..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Glutamicibacter (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Acidobacteria. (numeric)}{Abundance of this taxa/trait}
  \item{Haloferula (numeric)}{Abundance of this taxa/trait}
  \item{Thiomonas (numeric)}{Abundance of this taxa/trait}
  \item{Azohydromonas (numeric)}{Abundance of this taxa/trait}
  \item{Nocardiopsis (numeric)}{Abundance of this taxa/trait}
  \item{Limnohabitans (numeric)}{Abundance of this taxa/trait}
  \item{Rubrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Chlorobium (numeric)}{Abundance of this taxa/trait}
  \item{Caulobacter (numeric)}{Abundance of this taxa/trait}
  \item{Thermorudis (numeric)}{Abundance of this taxa/trait}
  \item{Micromonospora (numeric)}{Abundance of this taxa/trait}
  \item{Schleiferia (numeric)}{Abundance of this taxa/trait}
  \item{Prevotella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Scalindua (numeric)}{Abundance of this taxa/trait}
  \item{Saprospira (numeric)}{Abundance of this taxa/trait}
  \item{Chelativorans (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Deltaproteobacteria..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Gulosibacter (numeric)}{Abundance of this taxa/trait}
  \item{Acinetobacter (numeric)}{Abundance of this taxa/trait}
  \item{Magnetospirillum (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Planctomycetales. (numeric)}{Abundance of this taxa/trait}
  \item{Mariniradius (numeric)}{Abundance of this taxa/trait}
  \item{Weeksella (numeric)}{Abundance of this taxa/trait}
  \item{Actinobaculum (numeric)}{Abundance of this taxa/trait}
  \item{Planctopirus (numeric)}{Abundance of this taxa/trait}
  \item{Nitrosospira (numeric)}{Abundance of this taxa/trait}
  \item{Arenibacter (numeric)}{Abundance of this taxa/trait}
  \item{Thermoanaerobaculum (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Planctomycetes (numeric)}{Abundance of this taxa/trait}
  \item{Scytonema (numeric)}{Abundance of this taxa/trait}
  \item{Riemerella (numeric)}{Abundance of this taxa/trait}
  \item{Thermogemmatispora (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Koribacter (numeric)}{Abundance of this taxa/trait}
  \item{Alistipes (numeric)}{Abundance of this taxa/trait}
  \item{Filomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Cyclobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Synechococcus (numeric)}{Abundance of this taxa/trait}
  \item{Cutibacterium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Opitutaceae (numeric)}{Abundance of this taxa/trait}
  \item{Asinibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Gaetbulibacter (numeric)}{Abundance of this taxa/trait}
  \item{Cytophaga (numeric)}{Abundance of this taxa/trait}
  \item{Dyella (numeric)}{Abundance of this taxa/trait}
  \item{Rhodoferax (numeric)}{Abundance of this taxa/trait}
  \item{Olivibacter (numeric)}{Abundance of this taxa/trait}
  \item{Desulfobulbus (numeric)}{Abundance of this taxa/trait}
  \item{Delftia (numeric)}{Abundance of this taxa/trait}
  \item{Leptolyngbya (numeric)}{Abundance of this taxa/trait}
  \item{Thermomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Xenophilus (numeric)}{Abundance of this taxa/trait}
  \item{Rubinisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Bellilinea (numeric)}{Abundance of this taxa/trait}
  \item{Salinispora (numeric)}{Abundance of this taxa/trait}
  \item{Pseudopedobacter (numeric)}{Abundance of this taxa/trait}
  \item{Janibacter (numeric)}{Abundance of this taxa/trait}
  \item{Rufibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Bacteroidales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Ramlibacter (numeric)}{Abundance of this taxa/trait}
  \item{Truepera (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Geodermatophilaceae (numeric)}{Abundance of this taxa/trait}
  \item{Terrabacter (numeric)}{Abundance of this taxa/trait}
  \item{Chthonomonas (numeric)}{Abundance of this taxa/trait}
  \item{Agreia (numeric)}{Abundance of this taxa/trait}
  \item{Herbaspirillum (numeric)}{Abundance of this taxa/trait}
  \item{Ornatilinea (numeric)}{Abundance of this taxa/trait}
  \item{Austwickia (numeric)}{Abundance of this taxa/trait}
  \item{Lamprocystis (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Bacteroidetes..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Proteiniphilum (numeric)}{Abundance of this taxa/trait}
  \item{Pseudarthrobacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhizobiales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Methylobacter (numeric)}{Abundance of this taxa/trait}
  \item{Aureimonas (numeric)}{Abundance of this taxa/trait}
  \item{Mucilaginibacter (numeric)}{Abundance of this taxa/trait}
  \item{Chelatococcus (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Green.non.sulfur.bacteria. (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Cryomorphaceae (numeric)}{Abundance of this taxa/trait}
  \item{Aminobacter (numeric)}{Abundance of this taxa/trait}
  \item{Porphyromonas (numeric)}{Abundance of this taxa/trait}
  \item{Lunatimonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Brocadia (numeric)}{Abundance of this taxa/trait}
  \item{Thermanaerothrix (numeric)}{Abundance of this taxa/trait}
  \item{Pseudorhodobacter (numeric)}{Abundance of this taxa/trait}
  \item{Aeromonas (numeric)}{Abundance of this taxa/trait}
  \item{Rubritepida (numeric)}{Abundance of this taxa/trait}
  \item{Vibrio (numeric)}{Abundance of this taxa/trait}
  \item{Glaciibacter (numeric)}{Abundance of this taxa/trait}
  \item{Actinopolymorpha (numeric)}{Abundance of this taxa/trait}
  \item{Leptolinea (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Bacteroidetes.Chlorobi.group (numeric)}{Abundance of this taxa/trait}
  \item{Ensifer (numeric)}{Abundance of this taxa/trait}
  \item{Roseovarius (numeric)}{Abundance of this taxa/trait}
  \item{Phycisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Verminephrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Dysgonomonas (numeric)}{Abundance of this taxa/trait}
  \item{Tannerella (numeric)}{Abundance of this taxa/trait}
  \item{Aquabacterium (numeric)}{Abundance of this taxa/trait}
  \item{Cellulosimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Blastococcus (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Armatimonadetes (numeric)}{Abundance of this taxa/trait}
  \item{Treponema (numeric)}{Abundance of this taxa/trait}
  \item{Paludibacter (numeric)}{Abundance of this taxa/trait}
  \item{Nitrosococcus (numeric)}{Abundance of this taxa/trait}
  \item{Patulibacter (numeric)}{Abundance of this taxa/trait}
  \item{Leptonema (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Rokubacteria (numeric)}{Abundance of this taxa/trait}
  \item{Nevskia (numeric)}{Abundance of this taxa/trait}
  \item{Dermacoccus (numeric)}{Abundance of this taxa/trait}
  \item{Pseudaminobacter (numeric)}{Abundance of this taxa/trait}
  \item{Oerskovia (numeric)}{Abundance of this taxa/trait}
  \item{Flaviflexus (numeric)}{Abundance of this taxa/trait}
  \item{Promicromonospora (numeric)}{Abundance of this taxa/trait}
  \item{Hyalangium (numeric)}{Abundance of this taxa/trait}
  \item{Microscilla (numeric)}{Abundance of this taxa/trait}
  \item{Methylosinus (numeric)}{Abundance of this taxa/trait}
  \item{Ornithinimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Kineosporia (numeric)}{Abundance of this taxa/trait}
  \item{Trichodesmium (numeric)}{Abundance of this taxa/trait}
  \item{Aromatoleum (numeric)}{Abundance of this taxa/trait}
  \item{Thiohalocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Nonlabens (numeric)}{Abundance of this taxa/trait}
  \item{Ruania (numeric)}{Abundance of this taxa/trait}
  \item{Archangium (numeric)}{Abundance of this taxa/trait}
  \item{Aequorivita (numeric)}{Abundance of this taxa/trait}
  \item{Noviherbaspirillum (numeric)}{Abundance of this taxa/trait}
  \item{Flexithrix (numeric)}{Abundance of this taxa/trait}
  \item{Smithella (numeric)}{Abundance of this taxa/trait}
  \item{Sulfuricella (numeric)}{Abundance of this taxa/trait}
  \item{Paucibacter (numeric)}{Abundance of this taxa/trait}
  \item{Mucinivorans (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Protochlamydia (numeric)}{Abundance of this taxa/trait}
  \item{X.Polyangium..brachysporum (numeric)}{Abundance of this taxa/trait}
  \item{Flectobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Crocinitomix (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.Zixibacteria (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Acetobacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Sphaerotilus (numeric)}{Abundance of this taxa/trait}
  \item{Belliella (numeric)}{Abundance of this taxa/trait}
  \item{Georgenia (numeric)}{Abundance of this taxa/trait}
  \item{Parvibaculum (numeric)}{Abundance of this taxa/trait}
  \item{Rathayibacter (numeric)}{Abundance of this taxa/trait}
  \item{Brucella (numeric)}{Abundance of this taxa/trait}
  \item{Maribacter (numeric)}{Abundance of this taxa/trait}
  \item{Angustibacter (numeric)}{Abundance of this taxa/trait}
  \item{Lacinutrix (numeric)}{Abundance of this taxa/trait}
  \item{Sporichthya (numeric)}{Abundance of this taxa/trait}
  \item{Arcticibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Flammeovirgaceae (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Myxococcales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Enhygromyxa (numeric)}{Abundance of this taxa/trait}
  \item{Vulgatibacter (numeric)}{Abundance of this taxa/trait}
  \item{Pseudacidovorax (numeric)}{Abundance of this taxa/trait}
  \item{Methylomonas (numeric)}{Abundance of this taxa/trait}
  \item{Nesterenkonia (numeric)}{Abundance of this taxa/trait}
  \item{Gloeobacter (numeric)}{Abundance of this taxa/trait}
  \item{Anditalea (numeric)}{Abundance of this taxa/trait}
  \item{Halomonas (numeric)}{Abundance of this taxa/trait}
  \item{Galbibacter (numeric)}{Abundance of this taxa/trait}
  \item{Agrococcus (numeric)}{Abundance of this taxa/trait}
  \item{Microcystis (numeric)}{Abundance of this taxa/trait}
  \item{Curtobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Solimonas (numeric)}{Abundance of this taxa/trait}
  \item{Staphylococcus (numeric)}{Abundance of this taxa/trait}
  \item{Gelidibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.candidate.division.NC10 (numeric)}{Abundance of this taxa/trait}
  \item{Draconibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Pandoraea (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Methylococcaceae (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Methylomirabilis (numeric)}{Abundance of this taxa/trait}
  \item{Modestobacter (numeric)}{Abundance of this taxa/trait}
  \item{Actinomadura (numeric)}{Abundance of this taxa/trait}
  \item{Kutzneria (numeric)}{Abundance of this taxa/trait}
  \item{Thiothrix (numeric)}{Abundance of this taxa/trait}
  \item{Azospira (numeric)}{Abundance of this taxa/trait}
  \item{Anaerophaga (numeric)}{Abundance of this taxa/trait}
  \item{Sunxiuqinia (numeric)}{Abundance of this taxa/trait}
  \item{Methylocaldum (numeric)}{Abundance of this taxa/trait}
  \item{Melioribacter (numeric)}{Abundance of this taxa/trait}
  \item{Acidiphilium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Halieaceae (numeric)}{Abundance of this taxa/trait}
  \item{Saccharicrinis (numeric)}{Abundance of this taxa/trait}
  \item{Xylanimonas (numeric)}{Abundance of this taxa/trait}
  \item{Rhodothermus (numeric)}{Abundance of this taxa/trait}
  \item{Haematobacter (numeric)}{Abundance of this taxa/trait}
  \item{Methylocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Hyphomonas (numeric)}{Abundance of this taxa/trait}
  \item{Castellaniella (numeric)}{Abundance of this taxa/trait}
  \item{Dietzia (numeric)}{Abundance of this taxa/trait}
  \item{Oscillatoria (numeric)}{Abundance of this taxa/trait}
  \item{Streptosporangium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhodobacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Fibrisoma (numeric)}{Abundance of this taxa/trait}
  \item{Owenweeksia (numeric)}{Abundance of this taxa/trait}
  \item{Brevibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Methyloglobulus (numeric)}{Abundance of this taxa/trait}
  \item{Nonomuraea (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Gottesmanbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Marivirga (numeric)}{Abundance of this taxa/trait}
  \item{Eisenibacter (numeric)}{Abundance of this taxa/trait}
  \item{Marinobacter (numeric)}{Abundance of this taxa/trait}
  \item{Microbispora (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Moduliflexus (numeric)}{Abundance of this taxa/trait}
  \item{Azorhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Caldimonas (numeric)}{Abundance of this taxa/trait}
  \item{Thiorhodococcus (numeric)}{Abundance of this taxa/trait}
  \item{Cystobacter (numeric)}{Abundance of this taxa/trait}
  \item{Tolypothrix (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Flavobacteria. (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Solirubrobacterales (numeric)}{Abundance of this taxa/trait}
  \item{Echinicola (numeric)}{Abundance of this taxa/trait}
  \item{Salinivirga (numeric)}{Abundance of this taxa/trait}
  \item{Saccharomonospora (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Acidithiobacillales (numeric)}{Abundance of this taxa/trait}
  \item{Caenimonas (numeric)}{Abundance of this taxa/trait}
  \item{Psychroserpens (numeric)}{Abundance of this taxa/trait}
  \item{Propionimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Streptacidiphilus (numeric)}{Abundance of this taxa/trait}
  \item{Pleomorphomonas (numeric)}{Abundance of this taxa/trait}
  \item{Blastochloris (numeric)}{Abundance of this taxa/trait}
  \item{Beggiatoa (numeric)}{Abundance of this taxa/trait}
  \item{Nafulsella (numeric)}{Abundance of this taxa/trait}
  \item{Cellulophaga (numeric)}{Abundance of this taxa/trait}
  \item{Rhodospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Psychroflexus (numeric)}{Abundance of this taxa/trait}
  \item{Rugosibacter (numeric)}{Abundance of this taxa/trait}
  \item{Plesiocystis (numeric)}{Abundance of this taxa/trait}
  \item{Flexilinea (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Peptococcaceae (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Vecturithrix (numeric)}{Abundance of this taxa/trait}
  \item{Saccharothrix (numeric)}{Abundance of this taxa/trait}
  \item{Brachymonas (numeric)}{Abundance of this taxa/trait}
  \item{Aliihoeflea (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Burkholderiaceae (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.BRC1 (numeric)}{Abundance of this taxa/trait}
  \item{Alicyclobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Lysinimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Pseudoclavibacter (numeric)}{Abundance of this taxa/trait}
  \item{Leeuwenhoekiella (numeric)}{Abundance of this taxa/trait}
  \item{Parabacteroides (numeric)}{Abundance of this taxa/trait}
  \item{Flammeovirga (numeric)}{Abundance of this taxa/trait}
  \item{Sinomonas (numeric)}{Abundance of this taxa/trait}
  \item{Cyanothece (numeric)}{Abundance of this taxa/trait}
  \item{Acidithiobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Lactobacillus (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..clostridial.firmicutes. (numeric)}{Abundance of this taxa/trait}
  \item{Geodermatophilus (numeric)}{Abundance of this taxa/trait}
  \item{Methylomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Porphyromonadaceae (numeric)}{Abundance of this taxa/trait}
  \item{Indibacter (numeric)}{Abundance of this taxa/trait}
  \item{Azonexus (numeric)}{Abundance of this taxa/trait}
  \item{Diaphorobacter (numeric)}{Abundance of this taxa/trait}
  \item{Spirochaeta (numeric)}{Abundance of this taxa/trait}
  \item{Rubellimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Empedobacter (numeric)}{Abundance of this taxa/trait}
  \item{Desulfotomaculum (numeric)}{Abundance of this taxa/trait}
  \item{Polymorphum (numeric)}{Abundance of this taxa/trait}
  \item{Edaphobacter (numeric)}{Abundance of this taxa/trait}
  \item{Acidimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Mumia (numeric)}{Abundance of this taxa/trait}
  \item{Adhaeribacter (numeric)}{Abundance of this taxa/trait}
  \item{Limnochorda (numeric)}{Abundance of this taxa/trait}
  \item{Stigmatella (numeric)}{Abundance of this taxa/trait}
  \item{Labrenzia (numeric)}{Abundance of this taxa/trait}
  \item{Thermonema (numeric)}{Abundance of this taxa/trait}
  \item{Thermaerobacter (numeric)}{Abundance of this taxa/trait}
  \item{Terriglobus (numeric)}{Abundance of this taxa/trait}
  \item{Oligotropha (numeric)}{Abundance of this taxa/trait}
  \item{Martelella (numeric)}{Abundance of this taxa/trait}
  \item{Luteibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Clostridiales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Oceanibaculum (numeric)}{Abundance of this taxa/trait}
  \item{Acidisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Humibacter (numeric)}{Abundance of this taxa/trait}
  \item{Rhodomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Simplicispira (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.SR1 (numeric)}{Abundance of this taxa/trait}
  \item{Methyloferula (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Bacteroidetes. (numeric)}{Abundance of this taxa/trait}
  \item{Thiocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Calothrix (numeric)}{Abundance of this taxa/trait}
  \item{Methylococcus (numeric)}{Abundance of this taxa/trait}
  \item{Tenacibaculum (numeric)}{Abundance of this taxa/trait}
  \item{Roseivirga (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Candidatus.Dependentiae (numeric)}{Abundance of this taxa/trait}
  \item{Marinilabilia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Microgenomates.group (numeric)}{Abundance of this taxa/trait}
  \item{Caballeronia (numeric)}{Abundance of this taxa/trait}
  \item{Paeniglutamicibacter (numeric)}{Abundance of this taxa/trait}
  \item{Kibdelosporangium (numeric)}{Abundance of this taxa/trait}
  \item{Lechevalieria (numeric)}{Abundance of this taxa/trait}
  \item{Hugenholtzia (numeric)}{Abundance of this taxa/trait}
  \item{Corallococcus (numeric)}{Abundance of this taxa/trait}
  \item{Planktothrix (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Nitrospira (numeric)}{Abundance of this taxa/trait}
  \item{Azovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Ferrovum (numeric)}{Abundance of this taxa/trait}
  \item{Sanguibacter (numeric)}{Abundance of this taxa/trait}
  \item{Diplosphaera..Verrucomicrobia. (numeric)}{Abundance of this taxa/trait}
  \item{Rudanella (numeric)}{Abundance of this taxa/trait}
  \item{Desulfuromonas (numeric)}{Abundance of this taxa/trait}
  \item{Pseudoalteromonas (numeric)}{Abundance of this taxa/trait}
  \item{Prolixibacter (numeric)}{Abundance of this taxa/trait}
  \item{Chromobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Acidithrix (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Saccharibacteria.oral.taxon.TM7x (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhodocyclaceae (numeric)}{Abundance of this taxa/trait}
  \item{Kordia (numeric)}{Abundance of this taxa/trait}
  \item{Thioclava (numeric)}{Abundance of this taxa/trait}
  \item{Moorella (numeric)}{Abundance of this taxa/trait}
  \item{Chitinimonas (numeric)}{Abundance of this taxa/trait}
  \item{Deferrisoma (numeric)}{Abundance of this taxa/trait}
  \item{Syntrophobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ruminiclostridium (numeric)}{Abundance of this taxa/trait}
  \item{Asticcacaulis (numeric)}{Abundance of this taxa/trait}
  \item{Ruegeria (numeric)}{Abundance of this taxa/trait}
  \item{Cecembia (numeric)}{Abundance of this taxa/trait}
  \item{Nitriliruptor (numeric)}{Abundance of this taxa/trait}
  \item{Butyricimonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Symbiobacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Verrucomicrobiales (numeric)}{Abundance of this taxa/trait}
  \item{Cryptosporangium (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Dadabacteria (numeric)}{Abundance of this taxa/trait}
  \item{Acidobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Dactylosporangium (numeric)}{Abundance of this taxa/trait}
  \item{Nostoc (numeric)}{Abundance of this taxa/trait}
  \item{Actinotignum (numeric)}{Abundance of this taxa/trait}
  \item{Erythrobacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Promicromonosporaceae (numeric)}{Abundance of this taxa/trait}
  \item{Jonesia (numeric)}{Abundance of this taxa/trait}
  \item{Pusillimonas (numeric)}{Abundance of this taxa/trait}
  \item{Thermopetrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Rhodonellum (numeric)}{Abundance of this taxa/trait}
  \item{Frigoribacterium (numeric)}{Abundance of this taxa/trait}
  \item{Phyllobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Tepidiphilus (numeric)}{Abundance of this taxa/trait}
  \item{Duganella (numeric)}{Abundance of this taxa/trait}
  \item{Methylopila (numeric)}{Abundance of this taxa/trait}
  \item{Bernardetia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Candidatus.Latescibacteria (numeric)}{Abundance of this taxa/trait}
  \item{Tepidimonas (numeric)}{Abundance of this taxa/trait}
  \item{Derxia (numeric)}{Abundance of this taxa/trait}
  \item{Mangrovimonas (numeric)}{Abundance of this taxa/trait}
  \item{Dermatophilus (numeric)}{Abundance of this taxa/trait}
  \item{Leisingera (numeric)}{Abundance of this taxa/trait}
  \item{Nitritalea (numeric)}{Abundance of this taxa/trait}
  \item{Thiocystis (numeric)}{Abundance of this taxa/trait}
  \item{Thermus (numeric)}{Abundance of this taxa/trait}
  \item{Stigonemataceae (numeric)}{Abundance of this taxa/trait}
  \item{Thiomargarita (numeric)}{Abundance of this taxa/trait}
  \item{Clavibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Sphingobacteriaceae (numeric)}{Abundance of this taxa/trait}
  \item{Oleiagrimonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Kuenenia (numeric)}{Abundance of this taxa/trait}
  \item{Zobellia (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Magnetobacterium (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.TM7.genomosp..GTL1 (numeric)}{Abundance of this taxa/trait}
  \item{Caldithrix (numeric)}{Abundance of this taxa/trait}
  \item{Achromatium (numeric)}{Abundance of this taxa/trait}
  \item{Parachlamydia (numeric)}{Abundance of this taxa/trait}
  \item{Sedimenticola (numeric)}{Abundance of this taxa/trait}
  \item{Kitasatospora (numeric)}{Abundance of this taxa/trait}
  \item{Desulfomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Pseudogulbenkiania (numeric)}{Abundance of this taxa/trait}
  \item{Pseudanabaena (numeric)}{Abundance of this taxa/trait}
  \item{Cesiribacter (numeric)}{Abundance of this taxa/trait}
  \item{Tistrella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Acetothermum (numeric)}{Abundance of this taxa/trait}
  \item{Ferriphaselus (numeric)}{Abundance of this taxa/trait}
  \item{Pelobacter (numeric)}{Abundance of this taxa/trait}
  \item{Plantibacter (numeric)}{Abundance of this taxa/trait}
  \item{Gillisia (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Cloacimonas (numeric)}{Abundance of this taxa/trait}
  \item{Uliginosibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Fodinicurvata (numeric)}{Abundance of this taxa/trait}
  \item{Trueperella (numeric)}{Abundance of this taxa/trait}
  \item{Cellvibrio (numeric)}{Abundance of this taxa/trait}
  \item{Desulfococcus (numeric)}{Abundance of this taxa/trait}
  \item{Dehalococcoides (numeric)}{Abundance of this taxa/trait}
  \item{Micavibrio (numeric)}{Abundance of this taxa/trait}
  \item{Citricoccus (numeric)}{Abundance of this taxa/trait}
  \item{Beijerinckia (numeric)}{Abundance of this taxa/trait}
  \item{Jannaschia (numeric)}{Abundance of this taxa/trait}
  \item{Isoptericola (numeric)}{Abundance of this taxa/trait}
  \item{Thioflavicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Desulfatitalea (numeric)}{Abundance of this taxa/trait}
  \item{Granulicella (numeric)}{Abundance of this taxa/trait}
  \item{Neorhizobium (numeric)}{Abundance of this taxa/trait}
  \item{Estrella (numeric)}{Abundance of this taxa/trait}
  \item{Paenarthrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Enterococcus (numeric)}{Abundance of this taxa/trait}
  \item{Marinobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Nomurabacteria (numeric)}{Abundance of this taxa/trait}
  \item{Alcanivorax (numeric)}{Abundance of this taxa/trait}
  \item{Aquamicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Salinarimonas (numeric)}{Abundance of this taxa/trait}
  \item{Polaribacter (numeric)}{Abundance of this taxa/trait}
  \item{Methylotenera (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Rubidus (numeric)}{Abundance of this taxa/trait}
  \item{Desulfobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Thiorhodovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Escherichia (numeric)}{Abundance of this taxa/trait}
  \item{Shewanella (numeric)}{Abundance of this taxa/trait}
  \item{Sideroxydans (numeric)}{Abundance of this taxa/trait}
  \item{Pannonibacter (numeric)}{Abundance of this taxa/trait}
  \item{Dehalogenimonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Moranbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Caenispirillum (numeric)}{Abundance of this taxa/trait}
  \item{Fischerella (numeric)}{Abundance of this taxa/trait}
  \item{Acetobacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Thiotrichales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Cryobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Afifella (numeric)}{Abundance of this taxa/trait}
  \item{Bizionia (numeric)}{Abundance of this taxa/trait}
  \item{Hylemonella (numeric)}{Abundance of this taxa/trait}
  \item{Actinocatenispora (numeric)}{Abundance of this taxa/trait}
  \item{Alkaliflexus (numeric)}{Abundance of this taxa/trait}
  \item{Methylocella (numeric)}{Abundance of this taxa/trait}
  \item{Lutibaculum (numeric)}{Abundance of this taxa/trait}
  \item{Roseobacter (numeric)}{Abundance of this taxa/trait}
  \item{Herminiimonas (numeric)}{Abundance of this taxa/trait}
  \item{Hydrogenobacter (numeric)}{Abundance of this taxa/trait}
  \item{Silvibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Wenzhouxiangella (numeric)}{Abundance of this taxa/trait}
  \item{Pararhodospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Altererythrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Sulfitobacter (numeric)}{Abundance of this taxa/trait}
  \item{Marinifilum (numeric)}{Abundance of this taxa/trait}
  \item{Smaragdicoccus (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Jettenia (numeric)}{Abundance of this taxa/trait}
  \item{Klebsiella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Roizmanbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Lautropia (numeric)}{Abundance of this taxa/trait}
  \item{Ruminococcus (numeric)}{Abundance of this taxa/trait}
  \item{Neisseria (numeric)}{Abundance of this taxa/trait}
  \item{bacterium.SM23_57 (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Parcubacteria (numeric)}{Abundance of this taxa/trait}
  \item{Desulfomonile (numeric)}{Abundance of this taxa/trait}
  \item{Desulfosarcina (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Magasanikbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Demetria (numeric)}{Abundance of this taxa/trait}
  \item{Solirubrum (numeric)}{Abundance of this taxa/trait}
  \item{Herbiconiux (numeric)}{Abundance of this taxa/trait}
  \item{Saccharopolyspora (numeric)}{Abundance of this taxa/trait}
  \item{Terracidiphilus (numeric)}{Abundance of this taxa/trait}
  \item{Loktanella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Paracaedibacter (numeric)}{Abundance of this taxa/trait}
  \item{Fibrella (numeric)}{Abundance of this taxa/trait}
  \item{Hahella (numeric)}{Abundance of this taxa/trait}
  \item{Actinokineospora (numeric)}{Abundance of this taxa/trait}
  \item{Dermabacter (numeric)}{Abundance of this taxa/trait}
  \item{Haliea (numeric)}{Abundance of this taxa/trait}
  \item{Desulfatibacillum (numeric)}{Abundance of this taxa/trait}
  \item{Celeribacter (numeric)}{Abundance of this taxa/trait}
  \item{Timonella (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Gemmatimonadetes. (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Woesebacteria (numeric)}{Abundance of this taxa/trait}
  \item{Kineococcus (numeric)}{Abundance of this taxa/trait}
  \item{Acaryochloris (numeric)}{Abundance of this taxa/trait}
  \item{Starkeya (numeric)}{Abundance of this taxa/trait}
  \item{Leptospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Collimonas (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..gamma.proteobacteria. (numeric)}{Abundance of this taxa/trait}
  \item{Hydrocarboniphaga (numeric)}{Abundance of this taxa/trait}
  \item{Ancylobacter (numeric)}{Abundance of this taxa/trait}
  \item{Phaeospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Hamadaea (numeric)}{Abundance of this taxa/trait}
  \item{Polycyclovorans (numeric)}{Abundance of this taxa/trait}
  \item{Glycomyces (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Kaiserbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Desulfonatronum (numeric)}{Abundance of this taxa/trait}
  \item{Enterobacter (numeric)}{Abundance of this taxa/trait}
  \item{Oceanicola (numeric)}{Abundance of this taxa/trait}
  \item{Frateuria (numeric)}{Abundance of this taxa/trait}
  \item{Nitrococcus (numeric)}{Abundance of this taxa/trait}
  \item{Roseivivax (numeric)}{Abundance of this taxa/trait}
  \item{Tsukamurella (numeric)}{Abundance of this taxa/trait}
  \item{Methyloceanibacter (numeric)}{Abundance of this taxa/trait}
  \item{Thioploca (numeric)}{Abundance of this taxa/trait}
  \item{Arhodomonas (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.TA06 (numeric)}{Abundance of this taxa/trait}
  \item{Anabaena (numeric)}{Abundance of this taxa/trait}
  \item{Odoribacter (numeric)}{Abundance of this taxa/trait}
  \item{Symbiobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Photobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Nitrospina (numeric)}{Abundance of this taxa/trait}
  \item{Advenella (numeric)}{Abundance of this taxa/trait}
  \item{Desulfosporosinus (numeric)}{Abundance of this taxa/trait}
  \item{Devriesea (numeric)}{Abundance of this taxa/trait}
  \item{Microcoleus (numeric)}{Abundance of this taxa/trait}
  \item{Geopsychrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Lascolabacillus (numeric)}{Abundance of this taxa/trait}
  \item{Cucumibacter (numeric)}{Abundance of this taxa/trait}
  \item{Campylobacter (numeric)}{Abundance of this taxa/trait}
  \item{Eubacterium (numeric)}{Abundance of this taxa/trait}
  \item{Gryllotalpicola (numeric)}{Abundance of this taxa/trait}
  \item{Muricauda (numeric)}{Abundance of this taxa/trait}
  \item{Thalassospira (numeric)}{Abundance of this taxa/trait}
  \item{Amorphus (numeric)}{Abundance of this taxa/trait}
  \item{Azotobacter (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Dojkabacteria (numeric)}{Abundance of this taxa/trait}
  \item{Bacteriovorax (numeric)}{Abundance of this taxa/trait}
  \item{Alcaligenes (numeric)}{Abundance of this taxa/trait}
  \item{Chloroherpeton (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Flavobacteriaceae (numeric)}{Abundance of this taxa/trait}
  \item{Kytococcus (numeric)}{Abundance of this taxa/trait}
  \item{Chitinilyticum (numeric)}{Abundance of this taxa/trait}
  \item{Aurantimonas (numeric)}{Abundance of this taxa/trait}
  \item{Nitrospirillum (numeric)}{Abundance of this taxa/trait}
  \item{Serratia (numeric)}{Abundance of this taxa/trait}
  \item{Thermocrispum (numeric)}{Abundance of this taxa/trait}
  \item{Ectothiorhodospira (numeric)}{Abundance of this taxa/trait}
  \item{Lyngbya (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Corynebacteriaceae (numeric)}{Abundance of this taxa/trait}
  \item{Desulfitobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Magnetomorum (numeric)}{Abundance of this taxa/trait}
  \item{Thermithiobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Yersinia (numeric)}{Abundance of this taxa/trait}
  \item{Sciscionella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Tenderia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Acidimicrobiia (numeric)}{Abundance of this taxa/trait}
  \item{Actinomycetospora (numeric)}{Abundance of this taxa/trait}
  \item{Silanimonas (numeric)}{Abundance of this taxa/trait}
  \item{SAR324.cluster (numeric)}{Abundance of this taxa/trait}
  \item{Salmonella (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Coxiellaceae (numeric)}{Abundance of this taxa/trait}
  \item{Herbidospora (numeric)}{Abundance of this taxa/trait}
  \item{Kiloniella (numeric)}{Abundance of this taxa/trait}
  \item{Thiorhodospira (numeric)}{Abundance of this taxa/trait}
  \item{Collinsella (numeric)}{Abundance of this taxa/trait}
  \item{Methylophaga (numeric)}{Abundance of this taxa/trait}
  \item{Terracoccus (numeric)}{Abundance of this taxa/trait}
  \item{Nisaea (numeric)}{Abundance of this taxa/trait}
  \item{Synechocystis (numeric)}{Abundance of this taxa/trait}
  \item{Alloactinosynnema (numeric)}{Abundance of this taxa/trait}
  \item{Salipiger (numeric)}{Abundance of this taxa/trait}
  \item{Chitiniphilus (numeric)}{Abundance of this taxa/trait}
  \item{Acidothermus (numeric)}{Abundance of this taxa/trait}
  \item{Haemophilus (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Lachnospiraceae (numeric)}{Abundance of this taxa/trait}
  \item{Allochromatium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Intrasporangiaceae (numeric)}{Abundance of this taxa/trait}
  \item{Puniceibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Microbulbifer (numeric)}{Abundance of this taxa/trait}
  \item{Haematomicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Spongiibacter (numeric)}{Abundance of this taxa/trait}
  \item{Catenulispora (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Oxalobacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Syntrophorhabdus (numeric)}{Abundance of this taxa/trait}
  \item{Salinimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Thermophagus (numeric)}{Abundance of this taxa/trait}
  \item{Thalassobaculum (numeric)}{Abundance of this taxa/trait}
  \item{Lysinibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Magnetospira (numeric)}{Abundance of this taxa/trait}
  \item{Catelliglobosispora (numeric)}{Abundance of this taxa/trait}
  \item{Pseudochrobactrum (numeric)}{Abundance of this taxa/trait}
  \item{Rickettsia (numeric)}{Abundance of this taxa/trait}
  \item{Jejuia (numeric)}{Abundance of this taxa/trait}
  \item{Geobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Balneatrix (numeric)}{Abundance of this taxa/trait}
  \item{Acidocella (numeric)}{Abundance of this taxa/trait}
  \item{Sanguibacteroides (numeric)}{Abundance of this taxa/trait}
  \item{Defluviimonas (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhodanobacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Criblamydia (numeric)}{Abundance of this taxa/trait}
  \item{Sutterella (numeric)}{Abundance of this taxa/trait}
  \item{Pleurocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Helicobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ferrimicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Desulfotignum (numeric)}{Abundance of this taxa/trait}
  \item{Coleofasciculus (numeric)}{Abundance of this taxa/trait}
  \item{Zunongwangia (numeric)}{Abundance of this taxa/trait}
  \item{Francisella (numeric)}{Abundance of this taxa/trait}
  \item{Geoalkalibacter (numeric)}{Abundance of this taxa/trait}
  \item{Desulfatiglans (numeric)}{Abundance of this taxa/trait}
  \item{Niveispirillum (numeric)}{Abundance of this taxa/trait}
  \item{Desulfobacter (numeric)}{Abundance of this taxa/trait}
  \item{Neochlamydia (numeric)}{Abundance of this taxa/trait}
  \item{Meganema (numeric)}{Abundance of this taxa/trait}
  \item{Stappia (numeric)}{Abundance of this taxa/trait}
  \item{Methylohalobius (numeric)}{Abundance of this taxa/trait}
  \item{Mahella (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Oscillatoriales..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Pelagibaca (numeric)}{Abundance of this taxa/trait}
  \item{Komagataeibacter (numeric)}{Abundance of this taxa/trait}
  \item{Eudoraea (numeric)}{Abundance of this taxa/trait}
  \item{Coraliomargarita (numeric)}{Abundance of this taxa/trait}
  \item{Leeia (numeric)}{Abundance of this taxa/trait}
  \item{Pantoea (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Spirochaetes (numeric)}{Abundance of this taxa/trait}
  \item{Gallionella (numeric)}{Abundance of this taxa/trait}
  \item{Tepidicaulis (numeric)}{Abundance of this taxa/trait}
  \item{Catenuloplanes (numeric)}{Abundance of this taxa/trait}
  \item{Desulfatirhabdium (numeric)}{Abundance of this taxa/trait}
  \item{Sandarakinorhabdus (numeric)}{Abundance of this taxa/trait}
  \item{Pseudooceanicola (numeric)}{Abundance of this taxa/trait}
  \item{Geminocystis (numeric)}{Abundance of this taxa/trait}
  \item{Methylovulum (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Acetothermia. (numeric)}{Abundance of this taxa/trait}
  \item{Prauserella (numeric)}{Abundance of this taxa/trait}
  \item{Winogradskyella (numeric)}{Abundance of this taxa/trait}
  \item{Altibacter (numeric)}{Abundance of this taxa/trait}
  \item{Arcanobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Frondihabitans (numeric)}{Abundance of this taxa/trait}
  \item{Xanthomarina (numeric)}{Abundance of this taxa/trait}
  \item{Gramella (numeric)}{Abundance of this taxa/trait}
  \item{Arcobacter (numeric)}{Abundance of this taxa/trait}
  \item{Thermomonospora (numeric)}{Abundance of this taxa/trait}
  \item{Rhodovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Ornithobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Leuconostoc (numeric)}{Abundance of this taxa/trait}
  \item{Euryhalocaulis (numeric)}{Abundance of this taxa/trait}
  \item{Halotalea (numeric)}{Abundance of this taxa/trait}
  \item{Caedibacter (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Magnetoglobus (numeric)}{Abundance of this taxa/trait}
  \item{Pseudohongiella (numeric)}{Abundance of this taxa/trait}
  \item{Paenirhodobacter (numeric)}{Abundance of this taxa/trait}
  \item{Ponticoccus..Rhodobacteraceae. (numeric)}{Abundance of this taxa/trait}
  \item{Rheinheimera (numeric)}{Abundance of this taxa/trait}
  \item{Williamsia (numeric)}{Abundance of this taxa/trait}
  \item{Segniliparus (numeric)}{Abundance of this taxa/trait}
  \item{Sediminicola (numeric)}{Abundance of this taxa/trait}
  \item{Actinoalloteichus (numeric)}{Abundance of this taxa/trait}
  \item{Chlorobaculum (numeric)}{Abundance of this taxa/trait}
  \item{Gracilimonas (numeric)}{Abundance of this taxa/trait}
  \item{Ketogulonicigenium (numeric)}{Abundance of this taxa/trait}
  \item{Pelagibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Oceanithermus (numeric)}{Abundance of this taxa/trait}
  \item{Elstera (numeric)}{Abundance of this taxa/trait}
  \item{Thalassobacter (numeric)}{Abundance of this taxa/trait}
  \item{Tolumonas (numeric)}{Abundance of this taxa/trait}
  \item{Blautia (numeric)}{Abundance of this taxa/trait}
  \item{Thermoactinomyces (numeric)}{Abundance of this taxa/trait}
  \item{Rothia (numeric)}{Abundance of this taxa/trait}
  \item{Thiolapillus (numeric)}{Abundance of this taxa/trait}
  \item{Waddlia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Parachlamydiaceae (numeric)}{Abundance of this taxa/trait}
  \item{Teredinibacter (numeric)}{Abundance of this taxa/trait}
  \item{Granulibacter (numeric)}{Abundance of this taxa/trait}
  \item{Tamlana (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Kryptonium (numeric)}{Abundance of this taxa/trait}
  \item{OMG.group (numeric)}{Abundance of this taxa/trait}
  \item{Rubritalea (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..alpha.subdivision. (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Kryptobacter (numeric)}{Abundance of this taxa/trait}
  \item{Moorea (numeric)}{Abundance of this taxa/trait}
  \item{Salisaeta (numeric)}{Abundance of this taxa/trait}
  \item{Sterolibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Sulfurovum (numeric)}{Abundance of this taxa/trait}
  \item{Nodosilinea (numeric)}{Abundance of this taxa/trait}
  \item{Cyanobium (numeric)}{Abundance of this taxa/trait}
  \item{Yonghaparkia (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Uhrbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Salinisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Maritimibacter (numeric)}{Abundance of this taxa/trait}
  \item{Methylophilus (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Yanofskybacteria (numeric)}{Abundance of this taxa/trait}
  \item{Marinovum (numeric)}{Abundance of this taxa/trait}
  \item{Sphaerochaeta (numeric)}{Abundance of this taxa/trait}
  \item{Marinithermus (numeric)}{Abundance of this taxa/trait}
  \item{Octadecabacter (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.WOR.1 (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Rhodospirillaceae (numeric)}{Abundance of this taxa/trait}
  \item{Aneurinibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Persicobacter (numeric)}{Abundance of this taxa/trait}
  \item{Fusobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Dictyoglomus (numeric)}{Abundance of this taxa/trait}
  \item{bacterium.SM23_31 (numeric)}{Abundance of this taxa/trait}
  \item{Alkalilimnicola (numeric)}{Abundance of this taxa/trait}
  \item{Oceanicaulis (numeric)}{Abundance of this taxa/trait}
  \item{Lachnoclostridium (numeric)}{Abundance of this taxa/trait}
  \item{Algibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Spongiibacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Verrucomicrobia.subdivision.6 (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Candidatus.Peregrinibacteria (numeric)}{Abundance of this taxa/trait}
  \item{Salegentibacter (numeric)}{Abundance of this taxa/trait}
  \item{Thiohalorhabdus (numeric)}{Abundance of this taxa/trait}
  \item{Chlorogloeopsis (numeric)}{Abundance of this taxa/trait}
  \item{Heliobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Verrucosispora (numeric)}{Abundance of this taxa/trait}
  \item{Sulfobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Phaeobacter (numeric)}{Abundance of this taxa/trait}
  \item{Faecalibacterium (integer)}{Abundance of this taxa/trait}
  \item{Mesonia (numeric)}{Abundance of this taxa/trait}
  \item{Flaviramulus (numeric)}{Abundance of this taxa/trait}
  \item{Longispora (numeric)}{Abundance of this taxa/trait}
  \item{Porphyrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Actinosynnema (numeric)}{Abundance of this taxa/trait}
  \item{Arthrospira (numeric)}{Abundance of this taxa/trait}
  \item{Bergeyella (numeric)}{Abundance of this taxa/trait}
  \item{Desulforegula (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Porticoccaceae (numeric)}{Abundance of this taxa/trait}
  \item{Caldisericum (integer)}{Abundance of this taxa/trait}
  \item{Siansivirga (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Levybacteria (numeric)}{Abundance of this taxa/trait}
  \item{Gluconacetobacter (numeric)}{Abundance of this taxa/trait}
  \item{Salinibacter (numeric)}{Abundance of this taxa/trait}
  \item{Actinospica (numeric)}{Abundance of this taxa/trait}
  \item{Aminomonas (numeric)}{Abundance of this taxa/trait}
  \item{Aliterella (numeric)}{Abundance of this taxa/trait}
  \item{Thalassobius (numeric)}{Abundance of this taxa/trait}
  \item{Microvirgula (numeric)}{Abundance of this taxa/trait}
  \item{Marichromatium (numeric)}{Abundance of this taxa/trait}
  \item{Paraoerskovia (numeric)}{Abundance of this taxa/trait}
  \item{Citreicella (numeric)}{Abundance of this taxa/trait}
  \item{Methylogaea (numeric)}{Abundance of this taxa/trait}
  \item{Gloeocapsa (numeric)}{Abundance of this taxa/trait}
  \item{Methylobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Actinopolyspora (numeric)}{Abundance of this taxa/trait}
  \item{Lentisphaera (numeric)}{Abundance of this taxa/trait}
  \item{Planktothricoides (numeric)}{Abundance of this taxa/trait}
  \item{Cohnella (numeric)}{Abundance of this taxa/trait}
  \item{Mobiluncus (numeric)}{Abundance of this taxa/trait}
  \item{Xylophilus (numeric)}{Abundance of this taxa/trait}
  \item{Desulfobacca (numeric)}{Abundance of this taxa/trait}
  \item{Chamaesiphon (numeric)}{Abundance of this taxa/trait}
  \item{Gulbenkiania (numeric)}{Abundance of this taxa/trait}
  \item{Ahrensia (numeric)}{Abundance of this taxa/trait}
  \item{Amantichitinum (numeric)}{Abundance of this taxa/trait}
  \item{Novispirillum (numeric)}{Abundance of this taxa/trait}
  \item{Gallaecimonas (numeric)}{Abundance of this taxa/trait}
  \item{Mesoflavibacter (numeric)}{Abundance of this taxa/trait}
  \item{Algiphilus (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Daviesbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Mastigocladopsis (numeric)}{Abundance of this taxa/trait}
  \item{Methylacidiphilum (numeric)}{Abundance of this taxa/trait}
  \item{Roseibium (numeric)}{Abundance of this taxa/trait}
  \item{Desulfobacula (numeric)}{Abundance of this taxa/trait}
  \item{Cylindrospermum (numeric)}{Abundance of this taxa/trait}
  \item{Imtechella (numeric)}{Abundance of this taxa/trait}
  \item{Simkania (numeric)}{Abundance of this taxa/trait}
  \item{Halocynthiibacter (numeric)}{Abundance of this taxa/trait}
  \item{Prosthecobacter (numeric)}{Abundance of this taxa/trait}
  \item{Dokdonia (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Phaeomarinobacter (numeric)}{Abundance of this taxa/trait}
  \item{Methylomarinum (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Glomeribacter (numeric)}{Abundance of this taxa/trait}
  \item{Providencia (numeric)}{Abundance of this taxa/trait}
  \item{Thalassomonas (numeric)}{Abundance of this taxa/trait}
  \item{Geitlerinema (numeric)}{Abundance of this taxa/trait}
  \item{Paraglaciecola (numeric)}{Abundance of this taxa/trait}
  \item{Salinibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Allokutzneria (numeric)}{Abundance of this taxa/trait}
  \item{Limnobacter (numeric)}{Abundance of this taxa/trait}
  \item{Kangiella (numeric)}{Abundance of this taxa/trait}
  \item{Stackebrandtia (numeric)}{Abundance of this taxa/trait}
  \item{Litoreibacter (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Candidatus.Omnitrophica (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Sphingobacteria. (numeric)}{Abundance of this taxa/trait}
  \item{Andreprevotia (numeric)}{Abundance of this taxa/trait}
  \item{Hoyosella (numeric)}{Abundance of this taxa/trait}
  \item{Tomitella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Odyssella (numeric)}{Abundance of this taxa/trait}
  \item{Pectobacterium (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Desulfobulbaceae (numeric)}{Abundance of this taxa/trait}
  \item{Kamptonema (numeric)}{Abundance of this taxa/trait}
  \item{Butyrivibrio (numeric)}{Abundance of this taxa/trait}
  \item{Blastomonas (numeric)}{Abundance of this taxa/trait}
  \item{Tropicibacter (numeric)}{Abundance of this taxa/trait}
  \item{Syntrophus (numeric)}{Abundance of this taxa/trait}
  \item{Oscillibacter (integer)}{Abundance of this taxa/trait}
  \item{Pseudodonghicola (numeric)}{Abundance of this taxa/trait}
  \item{Akkermansia (numeric)}{Abundance of this taxa/trait}
  \item{Anaerobiospirillum (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.CPR2 (numeric)}{Abundance of this taxa/trait}
  \item{Laribacter (numeric)}{Abundance of this taxa/trait}
  \item{Acidihalobacter (numeric)}{Abundance of this taxa/trait}
  \item{Robiginitalea (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.CPR3 (numeric)}{Abundance of this taxa/trait}
  \item{Perlucidibaca (numeric)}{Abundance of this taxa/trait}
  \item{Lebetimonas (numeric)}{Abundance of this taxa/trait}
  \item{Tumebacillus (numeric)}{Abundance of this taxa/trait}
  \item{Xylella (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Xiphinematobacter (numeric)}{Abundance of this taxa/trait}
  \item{Sneathiella (numeric)}{Abundance of this taxa/trait}
  \item{Leucothrix (numeric)}{Abundance of this taxa/trait}
  \item{Endozoicomonas (numeric)}{Abundance of this taxa/trait}
  \item{Formosa (numeric)}{Abundance of this taxa/trait}
  \item{Magnetococcus (numeric)}{Abundance of this taxa/trait}
  \item{Dethiosulfatarculus (numeric)}{Abundance of this taxa/trait}
  \item{Pelotomaculum (numeric)}{Abundance of this taxa/trait}
  \item{Halothiobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Mycetocola (numeric)}{Abundance of this taxa/trait}
  \item{Thermodesulfatator (numeric)}{Abundance of this taxa/trait}
  \item{Desulfocurvus (numeric)}{Abundance of this taxa/trait}
  \item{Pelodictyon (numeric)}{Abundance of this taxa/trait}
  \item{Sedimentitalea (numeric)}{Abundance of this taxa/trait}
  \item{Sulfurospirillum (integer)}{Abundance of this taxa/trait}
  \item{Nitrincola (numeric)}{Abundance of this taxa/trait}
  \item{Idiomarina (numeric)}{Abundance of this taxa/trait}
  \item{Chitinispirillum (numeric)}{Abundance of this taxa/trait}
  \item{Lentzea (numeric)}{Abundance of this taxa/trait}
  \item{Balneola (numeric)}{Abundance of this taxa/trait}
  \item{Selenomonas (numeric)}{Abundance of this taxa/trait}
  \item{Thermoanaerobacter (numeric)}{Abundance of this taxa/trait}
  \item{Joostella (numeric)}{Abundance of this taxa/trait}
  \item{Shigella (numeric)}{Abundance of this taxa/trait}
  \item{Veillonella (integer)}{Abundance of this taxa/trait}
  \item{candidate.division.WWE3 (numeric)}{Abundance of this taxa/trait}
  \item{Catabacter (numeric)}{Abundance of this taxa/trait}
  \item{Subdoligranulum (integer)}{Abundance of this taxa/trait}
  \item{Mesotoga (numeric)}{Abundance of this taxa/trait}
  \item{Desulfovirgula (numeric)}{Abundance of this taxa/trait}
  \item{Peptoclostridium (numeric)}{Abundance of this taxa/trait}
  \item{Thermosynechococcus (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..delta.subdivision. (integer)}{Abundance of this taxa/trait}
  \item{Pseudohaliea (numeric)}{Abundance of this taxa/trait}
  \item{Mastigocoleus (numeric)}{Abundance of this taxa/trait}
  \item{Megamonas (integer)}{Abundance of this taxa/trait}
  \item{Anaerotruncus (numeric)}{Abundance of this taxa/trait}
  \item{Desulfocarbo (numeric)}{Abundance of this taxa/trait}
  \item{Coxiella (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Desulfobacterales (numeric)}{Abundance of this taxa/trait}
  \item{Pseudovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Actibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Pacificimonas (numeric)}{Abundance of this taxa/trait}
  \item{Lampropedia (numeric)}{Abundance of this taxa/trait}
  \item{Wenxinia (numeric)}{Abundance of this taxa/trait}
  \item{Lactococcus (numeric)}{Abundance of this taxa/trait}
  \item{Sagittula (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Caulobacteraceae (numeric)}{Abundance of this taxa/trait}
  \item{Holospora (numeric)}{Abundance of this taxa/trait}
  \item{Xenorhabdus (numeric)}{Abundance of this taxa/trait}
  \item{Megasphaera (integer)}{Abundance of this taxa/trait}
  \item{Halobacteriovorax (numeric)}{Abundance of this taxa/trait}
  \item{Acetobacterium (integer)}{Abundance of this taxa/trait}
  \item{Desulfurivibrio (numeric)}{Abundance of this taxa/trait}
  \item{Slackia (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Magnetoovum (numeric)}{Abundance of this taxa/trait}
  \item{Prochlorothrix (numeric)}{Abundance of this taxa/trait}
  \item{Kordiimonas (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.WOR.3 (integer)}{Abundance of this taxa/trait}
  \item{Flexibacter (numeric)}{Abundance of this taxa/trait}
  \item{Barnesiella (numeric)}{Abundance of this taxa/trait}
  \item{Terasakiella (numeric)}{Abundance of this taxa/trait}
  \item{Syntrophomonas (numeric)}{Abundance of this taxa/trait}
  \item{Varibaculum (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Proteobacteria (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..beta.subdivision. (numeric)}{Abundance of this taxa/trait}
  \item{Listeria (numeric)}{Abundance of this taxa/trait}
  \item{Leptotrichia (integer)}{Abundance of this taxa/trait}
  \item{Dickeya (numeric)}{Abundance of this taxa/trait}
  \item{Deefgea (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Berkelbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Dasania (numeric)}{Abundance of this taxa/trait}
  \item{Anaerovorax (integer)}{Abundance of this taxa/trait}
  \item{Turicibacter (numeric)}{Abundance of this taxa/trait}
  \item{Renibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Virgibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Zhouia (numeric)}{Abundance of this taxa/trait}
  \item{Dinoroseobacter (numeric)}{Abundance of this taxa/trait}
  \item{Desulfospira (numeric)}{Abundance of this taxa/trait}
  \item{Desulfarculus (numeric)}{Abundance of this taxa/trait}
  \item{Sediminibacter (numeric)}{Abundance of this taxa/trait}
  \item{Aliiroseovarius (numeric)}{Abundance of this taxa/trait}
  \item{Hapalosiphon (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Peribacter (integer)}{Abundance of this taxa/trait}
  \item{Desulfonatronovibrio (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Omnitrophus (integer)}{Abundance of this taxa/trait}
  \item{Roseburia (integer)}{Abundance of this taxa/trait}
  \item{Dorea (integer)}{Abundance of this taxa/trait}
  \item{Kyrpidia (numeric)}{Abundance of this taxa/trait}
  \item{Clostridioides (integer)}{Abundance of this taxa/trait}
  \item{Prochlorococcus (integer)}{Abundance of this taxa/trait}
  \item{Alteromonas (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Giovannonibacteria (numeric)}{Abundance of this taxa/trait}
  \item{Mariprofundus (numeric)}{Abundance of this taxa/trait}
  \item{Aminiphilus (integer)}{Abundance of this taxa/trait}
  \item{Aminobacterium (integer)}{Abundance of this taxa/trait}
  \item{Hirschia (numeric)}{Abundance of this taxa/trait}
  \item{Citrobacter (numeric)}{Abundance of this taxa/trait}
  \item{Xenococcus (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Curtissbacteria (numeric)}{Abundance of this taxa/trait}
  \item{candidate.division.KD3.62 (numeric)}{Abundance of this taxa/trait}
  \item{Sporosarcina (numeric)}{Abundance of this taxa/trait}
  \item{Dialister (integer)}{Abundance of this taxa/trait}
  \item{Croceicoccus (integer)}{Abundance of this taxa/trait}
  \item{Colwellia (integer)}{Abundance of this taxa/trait}
  \item{Eggerthella (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Myxobacteria. (numeric)}{Abundance of this taxa/trait}
  \item{Synergistes (integer)}{Abundance of this taxa/trait}
  \item{Hydrogenivirga (numeric)}{Abundance of this taxa/trait}
  \item{Apibacter (numeric)}{Abundance of this taxa/trait}
  \item{Paeniclostridium (integer)}{Abundance of this taxa/trait}
  \item{Effusibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Oxalobacter (numeric)}{Abundance of this taxa/trait}
  \item{Pelosinus (numeric)}{Abundance of this taxa/trait}
  \item{Methylovorus (numeric)}{Abundance of this taxa/trait}
  \item{Fulvimarina (numeric)}{Abundance of this taxa/trait}
  \item{Holdemanella (integer)}{Abundance of this taxa/trait}
  \item{Brachyspira (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Amesbacteria (integer)}{Abundance of this taxa/trait}
  \item{Ammonifex (numeric)}{Abundance of this taxa/trait}
  \item{Olleya (numeric)}{Abundance of this taxa/trait}
  \item{Parvularcula (numeric)}{Abundance of this taxa/trait}
  \item{Moraxella (numeric)}{Abundance of this taxa/trait}
  \item{Pseudoduganella (numeric)}{Abundance of this taxa/trait}
  \item{Vogesella (numeric)}{Abundance of this taxa/trait}
  \item{Thermosinus (numeric)}{Abundance of this taxa/trait}
  \item{Spirillospora (numeric)}{Abundance of this taxa/trait}
  \item{Thermobispora (numeric)}{Abundance of this taxa/trait}
  \item{Paludibacterium (numeric)}{Abundance of this taxa/trait}
  \item{Psychromonas (integer)}{Abundance of this taxa/trait}
  \item{Gilvimarinus (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Amoebophilus (numeric)}{Abundance of this taxa/trait}
  \item{Endomicrobium (integer)}{Abundance of this taxa/trait}
  \item{Gluconobacter (numeric)}{Abundance of this taxa/trait}
  \item{Kerstersia (numeric)}{Abundance of this taxa/trait}
  \item{Persephonella (numeric)}{Abundance of this taxa/trait}
  \item{Aquaspirillum (numeric)}{Abundance of this taxa/trait}
  \item{Neptunomonas (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Ruminococcaceae (integer)}{Abundance of this taxa/trait}
  \item{Flagellimonas (integer)}{Abundance of this taxa/trait}
  \item{Crocosphaera (integer)}{Abundance of this taxa/trait}
  \item{candidate.division.CPR1 (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Babela (integer)}{Abundance of this taxa/trait}
  \item{Thiomicrospira (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Collierbacteria (integer)}{Abundance of this taxa/trait}
  \item{Desulfocapsa (integer)}{Abundance of this taxa/trait}
  \item{Psychrobacter (integer)}{Abundance of this taxa/trait}
  \item{Catenibacterium (integer)}{Abundance of this taxa/trait}
  \item{bacterium.OL.1 (integer)}{Abundance of this taxa/trait}
  \item{Photorhabdus (numeric)}{Abundance of this taxa/trait}
  \item{Desulfohalobium (numeric)}{Abundance of this taxa/trait}
  \item{Aestuariivita (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Falkowbacteria (integer)}{Abundance of this taxa/trait}
  \item{Rivularia (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Candidatus.Peribacteria (integer)}{Abundance of this taxa/trait}
  \item{Bartonella (numeric)}{Abundance of this taxa/trait}
  \item{Sulfuricurvum (integer)}{Abundance of this taxa/trait}
  \item{Tatlockia (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Nocardioidaceae (numeric)}{Abundance of this taxa/trait}
  \item{Bilophila (integer)}{Abundance of this taxa/trait}
  \item{Olsenella (integer)}{Abundance of this taxa/trait}
  \item{Crinalium (numeric)}{Abundance of this taxa/trait}
  \item{Citromicrobium (numeric)}{Abundance of this taxa/trait}
  \item{Ferrimonas (integer)}{Abundance of this taxa/trait}
  \item{Chroococcidiopsis (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Clostridiaceae (numeric)}{Abundance of this taxa/trait}
  \item{Phascolarctobacterium (integer)}{Abundance of this taxa/trait}
  \item{Marinomonas (integer)}{Abundance of this taxa/trait}
  \item{Halothece (numeric)}{Abundance of this taxa/trait}
  \item{Asaia (numeric)}{Abundance of this taxa/trait}
  \item{Thermincola (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Shapirobacteria (integer)}{Abundance of this taxa/trait}
  \item{Chitinibacter (integer)}{Abundance of this taxa/trait}
  \item{Aphanizomenon (numeric)}{Abundance of this taxa/trait}
  \item{Terrisporobacter (integer)}{Abundance of this taxa/trait}
  \item{Elusimicrobium (integer)}{Abundance of this taxa/trait}
  \item{Polynucleobacter (numeric)}{Abundance of this taxa/trait}
  \item{endosymbiont.of.Ridgeia.piscesae (integer)}{Abundance of this taxa/trait}
  \item{Nereida (numeric)}{Abundance of this taxa/trait}
  \item{Pseudobacteroides (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Desulforudis (numeric)}{Abundance of this taxa/trait}
  \item{Thermotoga (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Pacebacteria (integer)}{Abundance of this taxa/trait}
  \item{Tatumella (numeric)}{Abundance of this taxa/trait}
  \item{Thermodesulfobacterium (integer)}{Abundance of this taxa/trait}
  \item{Solemya.velum.gill.symbiont (integer)}{Abundance of this taxa/trait}
  \item{Neosynechococcus (integer)}{Abundance of this taxa/trait}
  \item{Coprococcus (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Cyanobacteria..miscellaneous. (numeric)}{Abundance of this taxa/trait}
  \item{Erwinia (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Hyphomonadaceae (numeric)}{Abundance of this taxa/trait}
  \item{Cycloclasticus (integer)}{Abundance of this taxa/trait}
  \item{Myxosarcina (integer)}{Abundance of this taxa/trait}
  \item{Maricaulis (integer)}{Abundance of this taxa/trait}
  \item{Nautella (numeric)}{Abundance of this taxa/trait}
  \item{Aquitalea (integer)}{Abundance of this taxa/trait}
  \item{Cyanobacterium (numeric)}{Abundance of this taxa/trait}
  \item{Halorhodospira (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Verrucomicrobia. (integer)}{Abundance of this taxa/trait}
  \item{Marinagarivorans (integer)}{Abundance of this taxa/trait}
  \item{Salinispira (integer)}{Abundance of this taxa/trait}
  \item{Zooshikella (integer)}{Abundance of this taxa/trait}
  \item{Stanieria (integer)}{Abundance of this taxa/trait}
  \item{Simiduia (numeric)}{Abundance of this taxa/trait}
  \item{Phormidesmis (integer)}{Abundance of this taxa/trait}
  \item{Sediminispirochaeta (integer)}{Abundance of this taxa/trait}
  \item{Coprobacillus (numeric)}{Abundance of this taxa/trait}
  \item{Fibrobacter (integer)}{Abundance of this taxa/trait}
  \item{Thermodesulfovibrio (integer)}{Abundance of this taxa/trait}
  \item{Acidaminococcus (integer)}{Abundance of this taxa/trait}
  \item{Calditerricola (numeric)}{Abundance of this taxa/trait}
  \item{Candidatus.Jorgensenbacteria (integer)}{Abundance of this taxa/trait}
  \item{Marinimicrobium (integer)}{Abundance of this taxa/trait}
  \item{Spirulina (integer)}{Abundance of this taxa/trait}
  \item{marine.Actinobacteria.clade (integer)}{Abundance of this taxa/trait}
  \item{Saccharophagus (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Azambacteria (integer)}{Abundance of this taxa/trait}
  \item{Acholeplasma (integer)}{Abundance of this taxa/trait}
  \item{Orientia (integer)}{Abundance of this taxa/trait}
  \item{Desulfonatronospira (integer)}{Abundance of this taxa/trait}
  \item{Acetoanaerobium (integer)}{Abundance of this taxa/trait}
  \item{Carboxydothermus (numeric)}{Abundance of this taxa/trait}
  \item{Zymomonas (numeric)}{Abundance of this taxa/trait}
  \item{Hippea (integer)}{Abundance of this taxa/trait}
  \item{Luminiphilus (integer)}{Abundance of this taxa/trait}
  \item{Alkaliphilus (integer)}{Abundance of this taxa/trait}
  \item{Acetivibrio (integer)}{Abundance of this taxa/trait}
  \item{Ponticaulis (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Peptostreptococcaceae (integer)}{Abundance of this taxa/trait}
  \item{Caldicellulosiruptor (integer)}{Abundance of this taxa/trait}
  \item{Thermocrinis (integer)}{Abundance of this taxa/trait}
  \item{Mycoplasma (integer)}{Abundance of this taxa/trait}
  \item{Pseudoglutamicibacter (numeric)}{Abundance of this taxa/trait}
  \item{Intestinimonas (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Wolfebacteria (integer)}{Abundance of this taxa/trait}
  \item{Wolbachia (integer)}{Abundance of this taxa/trait}
  \item{Sporomusa (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Holosporaceae (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Chlamydiae (integer)}{Abundance of this taxa/trait}
  \item{Nannocystis (integer)}{Abundance of this taxa/trait}
  \item{Oceanospirillum (integer)}{Abundance of this taxa/trait}
  \item{Desmospora (integer)}{Abundance of this taxa/trait}
  \item{Calditerrivibrio (numeric)}{Abundance of this taxa/trait}
  \item{Congregibacter (integer)}{Abundance of this taxa/trait}
  \item{Acetonema (integer)}{Abundance of this taxa/trait}
  \item{Gynuella (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Flavobacteriia (integer)}{Abundance of this taxa/trait}
  \item{Fluoribacter (numeric)}{Abundance of this taxa/trait}
  \item{Acetomicrobium (integer)}{Abundance of this taxa/trait}
  \item{Anoxybacillus (integer)}{Abundance of this taxa/trait}
  \item{Mitsuokella (integer)}{Abundance of this taxa/trait}
  \item{Croceibacter (integer)}{Abundance of this taxa/trait}
  \item{Hellea (integer)}{Abundance of this taxa/trait}
  \item{Desulfofustis (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Actinobacteria. (integer)}{Abundance of this taxa/trait}
  \item{Thermacetogenium (integer)}{Abundance of this taxa/trait}
  \item{Oceaniovalibus (numeric)}{Abundance of this taxa/trait}
  \item{Yaniella (numeric)}{Abundance of this taxa/trait}
  \item{Robinsoniella (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Cytophagales. (numeric)}{Abundance of this taxa/trait}
  \item{Desulfitibacter (integer)}{Abundance of this taxa/trait}
  \item{Sulfurimonas (integer)}{Abundance of this taxa/trait}
  \item{Glaciecola (integer)}{Abundance of this taxa/trait}
  \item{Prosthecochloris (integer)}{Abundance of this taxa/trait}
  \item{Shimia (numeric)}{Abundance of this taxa/trait}
  \item{Mesoaciditoga (numeric)}{Abundance of this taxa/trait}
  \item{environmental.samples..Acidobacteriales. (integer)}{Abundance of this taxa/trait}
  \item{Rubidibacter (integer)}{Abundance of this taxa/trait}
  \item{Romboutsia (integer)}{Abundance of this taxa/trait}
  \item{Henriciella (integer)}{Abundance of this taxa/trait}
  \item{Limnoraphis (integer)}{Abundance of this taxa/trait}
  \item{Youngiibacter (integer)}{Abundance of this taxa/trait}
  \item{Pseudoflavonifractor (integer)}{Abundance of this taxa/trait}
  \item{Woodsholea (integer)}{Abundance of this taxa/trait}
  \item{Maribius (numeric)}{Abundance of this taxa/trait}
  \item{Alishewanella (integer)}{Abundance of this taxa/trait}
  \item{Cardiobacterium (integer)}{Abundance of this taxa/trait}
  \item{Oceanobacillus (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Campbellbacteria (numeric)}{Abundance of this taxa/trait}
  \item{Ochrovirga (integer)}{Abundance of this taxa/trait}
  \item{Thermanaerovibrio (integer)}{Abundance of this taxa/trait}
  \item{Dactylococcopsis (integer)}{Abundance of this taxa/trait}
  \item{Halanaerobium (integer)}{Abundance of this taxa/trait}
  \item{Acidomonas (numeric)}{Abundance of this taxa/trait}
  \item{Desulfotalea (integer)}{Abundance of this taxa/trait}
  \item{Desulfovermiculus (integer)}{Abundance of this taxa/trait}
  \item{Saccharospirillum (integer)}{Abundance of this taxa/trait}
  \item{Thermovirga (integer)}{Abundance of this taxa/trait}
  \item{Chrysiogenes (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Flavobacteriales. (numeric)}{Abundance of this taxa/trait}
  \item{Thalassotalea (integer)}{Abundance of this taxa/trait}
  \item{Coprothermobacter (integer)}{Abundance of this taxa/trait}
  \item{Kosmotoga (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Beckwithbacteria (integer)}{Abundance of this taxa/trait}
  \item{Catenovulum (integer)}{Abundance of this taxa/trait}
  \item{Moritella (integer)}{Abundance of this taxa/trait}
  \item{Amphritea (integer)}{Abundance of this taxa/trait}
  \item{Peptoniphilus (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Puniceispirillum (integer)}{Abundance of this taxa/trait}
  \item{Dolichospermum (integer)}{Abundance of this taxa/trait}
  \item{Geovibrio (integer)}{Abundance of this taxa/trait}
  \item{Nitratifractor (integer)}{Abundance of this taxa/trait}
  \item{Thioalkalimicrobium (integer)}{Abundance of this taxa/trait}
  \item{Hungatella (integer)}{Abundance of this taxa/trait}
  \item{Thermoanaerobacterium (integer)}{Abundance of this taxa/trait}
  \item{Kingella (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Alteromonadaceae (integer)}{Abundance of this taxa/trait}
  \item{Senegalimassilia (integer)}{Abundance of this taxa/trait}
  \item{Thermosulfidibacter (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Jidaibacter (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Oceanospirillaceae (integer)}{Abundance of this taxa/trait}
  \item{Pseudoramibacter (integer)}{Abundance of this taxa/trait}
  \item{Sulfurihydrogenibium (integer)}{Abundance of this taxa/trait}
  \item{Arenitalea (integer)}{Abundance of this taxa/trait}
  \item{Croceitalea (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Rickettsiales..miscellaneous. (integer)}{Abundance of this taxa/trait}
  \item{Oleispira (integer)}{Abundance of this taxa/trait}
  \item{Caldicoprobacter (integer)}{Abundance of this taxa/trait}
  \item{Robiginitomaculum (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Arcanobacter (integer)}{Abundance of this taxa/trait}
  \item{Aliagarivorans (integer)}{Abundance of this taxa/trait}
  \item{Dethiobacter (integer)}{Abundance of this taxa/trait}
  \item{Syntrophothermus (integer)}{Abundance of this taxa/trait}
  \item{Syntrophaceticus (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Bradyrhizobiaceae (integer)}{Abundance of this taxa/trait}
  \item{Enhydrobacter (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Sphingomonadaceae (integer)}{Abundance of this taxa/trait}
  \item{Alkanindiges (integer)}{Abundance of this taxa/trait}
  \item{Dethiosulfovibrio (integer)}{Abundance of this taxa/trait}
  \item{Phormidium (integer)}{Abundance of this taxa/trait}
  \item{Gorillibacterium (integer)}{Abundance of this taxa/trait}
  \item{Oxobacter (integer)}{Abundance of this taxa/trait}
  \item{Paraprevotella (integer)}{Abundance of this taxa/trait}
  \item{Atopobium (integer)}{Abundance of this taxa/trait}
  \item{Maritalea (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Methylopumilus (integer)}{Abundance of this taxa/trait}
  \item{Asanoa (numeric)}{Abundance of this taxa/trait}
  \item{Mastigocladus (integer)}{Abundance of this taxa/trait}
  \item{Aerococcaceae (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Pelagibacter (integer)}{Abundance of this taxa/trait}
  \item{Vermiphilus (integer)}{Abundance of this taxa/trait}
  \item{Desulfurispirillum (integer)}{Abundance of this taxa/trait}
  \item{Plesiomonas (integer)}{Abundance of this taxa/trait}
  \item{Thalassolituus (integer)}{Abundance of this taxa/trait}
  \item{Piscirickettsia (integer)}{Abundance of this taxa/trait}
  \item{Fictibacillus (numeric)}{Abundance of this taxa/trait}
  \item{Pontibacillus (integer)}{Abundance of this taxa/trait}
  \item{SAR86.cluster (integer)}{Abundance of this taxa/trait}
  \item{Rikenella (integer)}{Abundance of this taxa/trait}
  \item{Sebaldella (integer)}{Abundance of this taxa/trait}
  \item{Aliiglaciecola (integer)}{Abundance of this taxa/trait}
  \item{Algicola (integer)}{Abundance of this taxa/trait}
  \item{Reinekea (integer)}{Abundance of this taxa/trait}
  \item{Orenia (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Clostridia (integer)}{Abundance of this taxa/trait}
  \item{Deferribacter (integer)}{Abundance of this taxa/trait}
  \item{Oligella (integer)}{Abundance of this taxa/trait}
  \item{Vitreoscilla (integer)}{Abundance of this taxa/trait}
  \item{Rickettsiella (integer)}{Abundance of this taxa/trait}
  \item{Allosalinactinospora (numeric)}{Abundance of this taxa/trait}
  \item{Planomicrobium (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Izimaplasma (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Rhizobiaceae.group. (integer)}{Abundance of this taxa/trait}
  \item{Zhongshania (integer)}{Abundance of this taxa/trait}
  \item{Propionispora (integer)}{Abundance of this taxa/trait}
  \item{Marinitoga (integer)}{Abundance of this taxa/trait}
  \item{Pseudothermotoga (integer)}{Abundance of this taxa/trait}
  \item{Alloprevotella (integer)}{Abundance of this taxa/trait}
  \item{Oceanimonas (integer)}{Abundance of this taxa/trait}
  \item{Spiribacter (numeric)}{Abundance of this taxa/trait}
  \item{Thermobifida (integer)}{Abundance of this taxa/trait}
  \item{Oribacterium (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Enterobacteriaceae (integer)}{Abundance of this taxa/trait}
  \item{Sporolactobacillus (integer)}{Abundance of this taxa/trait}
  \item{Carnobacterium (integer)}{Abundance of this taxa/trait}
  \item{Caloranaerobacter (integer)}{Abundance of this taxa/trait}
  \item{Halothermothrix (integer)}{Abundance of this taxa/trait}
  \item{Anaeromusa (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Gemmatimonadales. (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Alteromonadales (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Thioglobus (integer)}{Abundance of this taxa/trait}
  \item{Gracilibacillus (integer)}{Abundance of this taxa/trait}
  \item{Cellulosilyticum (integer)}{Abundance of this taxa/trait}
  \item{Halobacteroides (integer)}{Abundance of this taxa/trait}
  \item{Caldanaerobius (integer)}{Abundance of this taxa/trait}
  \item{Agarivorans (integer)}{Abundance of this taxa/trait}
  \item{Lacimicrobium (integer)}{Abundance of this taxa/trait}
  \item{Intestinibacter (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Methylophilaceae (integer)}{Abundance of this taxa/trait}
  \item{endosymbiont.of.Riftia.pachyptila (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Verrucomicrobiales. (integer)}{Abundance of this taxa/trait}
  \item{Rhodoluna (integer)}{Abundance of this taxa/trait}
  \item{Domibacillus (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Erysipelotrichaceae (integer)}{Abundance of this taxa/trait}
  \item{Chitinivibrio (integer)}{Abundance of this taxa/trait}
  \item{Arsukibacterium (integer)}{Abundance of this taxa/trait}
  \item{Anaerostipes (integer)}{Abundance of this taxa/trait}
  \item{Porticoccus (integer)}{Abundance of this taxa/trait}
  \item{Halobacillus (integer)}{Abundance of this taxa/trait}
  \item{Saccharibacillus (integer)}{Abundance of this taxa/trait}
  \item{Desnuesiella (integer)}{Abundance of this taxa/trait}
  \item{bacterium.MS4 (integer)}{Abundance of this taxa/trait}
  \item{Desulfurispora (integer)}{Abundance of this taxa/trait}
  \item{Taylorella (integer)}{Abundance of this taxa/trait}
  \item{Nodularia (integer)}{Abundance of this taxa/trait}
  \item{Fortiea (integer)}{Abundance of this taxa/trait}
  \item{Acetohalobium (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Pelagibacteraceae (integer)}{Abundance of this taxa/trait}
  \item{Gordonibacter (integer)}{Abundance of this taxa/trait}
  \item{Weissella (integer)}{Abundance of this taxa/trait}
  \item{Desulfurobacterium (integer)}{Abundance of this taxa/trait}
  \item{Planktomarina (integer)}{Abundance of this taxa/trait}
  \item{ac1.cluster (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Proteobacteria. (integer)}{Abundance of this taxa/trait}
  \item{Natranaerobius (integer)}{Abundance of this taxa/trait}
  \item{Edwardsiella (integer)}{Abundance of this taxa/trait}
  \item{Falsirhodobacter (integer)}{Abundance of this taxa/trait}
  \item{Neptuniibacter (integer)}{Abundance of this taxa/trait}
  \item{Butyricicoccus (integer)}{Abundance of this taxa/trait}
  \item{Aestuariibacter (integer)}{Abundance of this taxa/trait}
  \item{Herbinix (integer)}{Abundance of this taxa/trait}
  \item{Gayadomonas (integer)}{Abundance of this taxa/trait}
  \item{Bermanella (integer)}{Abundance of this taxa/trait}
  \item{Brenneria (numeric)}{Abundance of this taxa/trait}
  \item{Streptomonospora (integer)}{Abundance of this taxa/trait}
  \item{Caloramator (integer)}{Abundance of this taxa/trait}
  \item{Gottschalkia (integer)}{Abundance of this taxa/trait}
  \item{Confluentimicrobium (integer)}{Abundance of this taxa/trait}
  \item{Occidentia (integer)}{Abundance of this taxa/trait}
  \item{Oceanobacter (integer)}{Abundance of this taxa/trait}
  \item{Gallibacterium (integer)}{Abundance of this taxa/trait}
  \item{Acaricomes (numeric)}{Abundance of this taxa/trait}
  \item{Ornithinibacillus (integer)}{Abundance of this taxa/trait}
  \item{Anaeromassilibacillus (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Firmicutes.sensu.stricto..miscellaneous. (integer)}{Abundance of this taxa/trait}
  \item{Fervidobacterium (integer)}{Abundance of this taxa/trait}
  \item{Flexistipes (integer)}{Abundance of this taxa/trait}
  \item{Nesiotobacter (integer)}{Abundance of this taxa/trait}
  \item{Fangia (integer)}{Abundance of this taxa/trait}
  \item{Cylindrospermopsis (integer)}{Abundance of this taxa/trait}
  \item{Raphidiopsis (integer)}{Abundance of this taxa/trait}
  \item{Thermicanus (integer)}{Abundance of this taxa/trait}
  \item{Paenisporosarcina (integer)}{Abundance of this taxa/trait}
  \item{Risungbinella (integer)}{Abundance of this taxa/trait}
  \item{Geomicrobium (integer)}{Abundance of this taxa/trait}
  \item{X.Eubacterium..brachy (integer)}{Abundance of this taxa/trait}
  \item{Tepidanaerobacter (integer)}{Abundance of this taxa/trait}
  \item{Blattabacterium (integer)}{Abundance of this taxa/trait}
  \item{Cronobacter (integer)}{Abundance of this taxa/trait}
  \item{Morganella (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Zetaproteobacteria (integer)}{Abundance of this taxa/trait}
  \item{Planococcus (integer)}{Abundance of this taxa/trait}
  \item{X.Eubacterium..nodatum (integer)}{Abundance of this taxa/trait}
  \item{X.Clostridium..ultunense (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Erythrobacteraceae (integer)}{Abundance of this taxa/trait}
  \item{environmental.samples..Desulfobacterales. (integer)}{Abundance of this taxa/trait}
  \item{Aliivibrio (integer)}{Abundance of this taxa/trait}
  \item{Grimontia (integer)}{Abundance of this taxa/trait}
  \item{Borreliaceae (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Microbacteriaceae (integer)}{Abundance of this taxa/trait}
  \item{Thermobacillus (integer)}{Abundance of this taxa/trait}
  \item{Jeotgalibacillus (integer)}{Abundance of this taxa/trait}
  \item{Pediococcus (integer)}{Abundance of this taxa/trait}
  \item{Fervidicella (integer)}{Abundance of this taxa/trait}
  \item{Mogibacterium (integer)}{Abundance of this taxa/trait}
  \item{Niameybacter (integer)}{Abundance of this taxa/trait}
  \item{Candidatus.Soleaferrea (integer)}{Abundance of this taxa/trait}
  \item{Propionispira (integer)}{Abundance of this taxa/trait}
  \item{Rickettsiales.genera.incertae.sedis (numeric)}{Abundance of this taxa/trait}
  \item{unclassified.Mollicutes..miscellaneous. (integer)}{Abundance of this taxa/trait}
  \item{unclassified.Rickettsiaceae (integer)}{Abundance of this taxa/trait}
}
}
\source{
Experimental data from 3CA perturbation experiment. Santillan, Ezequiel, et al. "Traitbased lifehistory strategies explain succession scenario for complex bacterial communities under varying disturbance." Environmental microbiology 21.10 (2019): 3751-3764.
}
\usage{
CSR_TAXA_data
}
\description{
This dataset contains the following columns:
}
\keyword{datasets}
