% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{Hill_diversity}
\alias{Hill_diversity}
\title{Hill diversity analysis (MicroEcoTools)}
\usage{
Hill_diversity(dAtA)
}
\arguments{
\item{dAtA}{Dataframe is the count table.}
}
\value{
This function returns a table containing the Hill diversity values.
}
\description{
This function returns zero-, first- and second-order Hill diversity
}
\details{
The input data sample:\tabular{rrr}{
   TAXA1 \tab TAXA2 \tab TAXA3 \cr
   10 \tab 91 \tab 68 \cr
   11 \tab 86 \tab 70 \cr
   12 \tab 84 \tab 70 \cr
   15 \tab 30 \tab 3452 \cr
   19 \tab 45 \tab 3274 \cr
   13 \tab 25 \tab 3601 \cr
}
}
\examples{
Hill_diversity(dAtA = CSR_TAXA_data[3:1509])
}
