% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{NMA}
\alias{NMA}
\title{Theoretical Microbial Ecology Tools (MicroEcoTools)}
\usage{
NMA(DaTa, NSim, InDeX, ObSsIm, p.adj, Plot_Level, Keep_Data)
}
\arguments{
\item{DaTa}{A data frame containing the environmental and community data in long format with column names of Exp.Grp, Reactor, Time_point, TAXA, and Count Please note that you need at least two replicates for each experimental group.}

\item{NSim}{A number to specify the number of simulations (values bellow 1000 are not recomended). Default to 1000 simulations.}

\item{InDeX}{A diversity index from vegan package ("shannon", "simpson" or "invsimpson"). Default to "invsimpson.}

\item{ObSsIm}{A dataset with multinomial distribution using the observed dataset as probability matrix will be generated.}

\item{p.adj}{A p-value correction method ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Default to "Benjamini & Hochberg".}

\item{Plot_level}{A vector with sites/experimental group titles. Default to the unique values in experimental group column in your input data frame.}

\item{Keep_data}{A TRUE/FALSE variable to specify if you want to keep/remove simulated data. Default to FALSE.}
}
\value{
A list named NMA_Results containing the data (observed and simulated communities), results of the statistical analysis on the diversity metric between observed and simulated communities, calculated standard effect sizes and stochastic intensities, standard effect size and stochastic intensity plots using the user specified diversity metric (InDeX parameter), standard effect size and stochastic intensity plots using richness (based on Kraft et al. and Santillan et al. papers), and standard effect size plot using Cohen's d method.
}
\description{
Using a null model approach, this package tests the relative contribution of the assembly mechanisms (stochastic and deterministic) in a given community. Based on the observed communities, assuming that the community structure was randomly shaped, using a multinomial distribution, this package simulates random communities. A diversity metric will then be used to calculate the standard effect size between the observed communities and randomly generated communities.
}
\details{
The input data sample:\tabular{lcrcr}{
   Exp.Grp \tab Reactor \tab Time_point \tab TAXA \tab Count \cr
   I \tab 1 \tab 1 \tab rk1 \tab 68 \cr
   I \tab 2 \tab 1 \tab rk1 \tab 70 \cr
   X0 \tab 1 \tab 1 \tab rk1 \tab 3452 \cr
   X0 \tab 3 \tab 1 \tab rk1 \tab 3274 \cr
   X0 \tab 3 \tab 1 \tab rk1 \tab 3601 \cr
}
}
\note{
All environmental variables should be converted to factor

Environmental variables accepted are Time_point, Exp.Grp, and Reactor
}
\examples{
NMA(DaTa = NMA_data)

NMA(DaTa = NMA_data, NSim = 100, InDeX = "invsimpson", ObSsIm = FALSE, p.adj = "BH", Plot_Level = c("L0", "L1", "L2", "L3", "L4", "L5", "L6", "L7"), Keep_Data = FALSE)

}
\references{
Santillan, Ezequiel, et al. "Frequency of disturbance alters diversity, function, and underlying assembly mechanisms of complex bacterial communities." npj Biofilms and Microbiomes 5.1 (2019): 1-9.

Kraft, Nathan JB, et al. "Disentangling the drivers of β diversity along latitudinal and elevational gradients." Science 333.6050 (2011): 1755-1758.
}
\author{
Soheil A. Neshat, Ezequiel Santillan, Stefan Wuertz

(The NMA package was developed in Singapore Centre for Environmental Life Sciences Engineering, Nanyang Technological University, Singapore)
}
\keyword{Null}
\keyword{analysis,}
\keyword{community}
\keyword{ecology}
\keyword{model}
