% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{NMA_data}
\alias{NMA_data}
\title{NMA_data dataset}
\format{
A data frame with 24 rows and 1000+ columns:
\describe{
  \item{Exp.Grp}{Experimental group (factor).}
  \item{Replicate}{Replicate number (factor).}
  \item{Time_point}{Time point (factor).}
  \item{rk1, rk2, ..., rk1507}{Counts for different taxa (integers).}
}
}
\usage{
NMA_data
}
\description{
A dataset containing microbial taxa abundance across different experimental groups, replicates, and time points.
}
\details{
This dataset contains the abundance of microbial taxa for different experimental groups, replicates, and time points. The columns `rk1` to `rk999` represent different microbial taxa, and their values indicate the count of each taxon in the respective sample.
}
\examples{
data(NMA_data)
head(NMA_data)
}
\keyword{datasets}
