% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{NMA_plot}
\alias{NMA_plot}
\title{Plotting function for the Null Model Analysis (MicroEcoTools)}
\usage{
NMA_plot(
  NMA_Results,
  plot_type = c("SES", "SI", "SES_species_number", "SI_species_number", "Cohens_d",
    "Hill")
)
}
\arguments{
\item{NMA_Results}{A list returned by the NMA function.}

\item{plot_type}{A character string specifying the desired plot type. Valid options are:
\itemize{
  \item \code{"SES"}: Standard effect size plot using the specified diversity index.
  \item \code{"SI"}: Stochastic intensity plot using the specified diversity index.
  \item \code{"SES_species_number"}: Standard effect size plot using species richness as diversity measure.
  \item \code{"SI_species_number"}: Stochastic intensity plot using species richness as diversity measure.
  \item \code{"Cohens_d"}: Standard effect size plot based on Cohen's d.
  \item \code{"Hill"}: A list of plots for zero-, first-, and second-order Hill diversity.
}
Defaults to \code{"SES"}.}
}
\value{
A ggplot object corresponding to the chosen plot type. If \code{plot_type = "Hill"}, a list of ggplot objects is returned.
}
\description{
This function takes in the NMA_Results list and plots the requested plot by the user.
}
\examples{
\dontrun{
  # Perform null model analysis on experimental data from a perturbation experiment.
  nma_results <- NMA(DaTa = NMA_data, NSim = 1000)

  # Generate and display the sorted abundance plot:
  NMA_plot(nma_results, plot_type = "SES")

  # Generate and display the sorted adjusted P-value plot:
  NMA_plot(nma_results, plot_type = "SES_species_number")
}
}
