% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{pairwise_welch}
\alias{pairwise_welch}
\title{Pairwise Welch-ANOVA (MicroEcoTools)}
\usage{
pairwise_welch(dAtA, var.name, p_adj, v.equal, p.value.cutoff, Parallel)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}

\item{v.equal}{Assumption of equal variance can be forced using this parameter. By default it is set to FALSE where it cannot be set to FALSE it will switch to TRUE automatically; a note will be shown in the remarks column in those cases.}

\item{p.value.cutoff}{A cut-off value for calling the P-value significant can be set using this parameter. The default value is 0.05.}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations.}
}
\value{
This function returns a table containing the pairwise statistical comparison results. The output table can be fed into the CSR_assign function to assign CSR categories.
}
\description{
Pairwise comparison between experimental groups using Welch-ANOVA or ANOVA. This function uses oneway test assuming non-equal variance to perform pairwise comparison.
In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE it can be forced to use equal/not equal variance by setting v.equal to TRUE or FALSE.
}
\details{
The input data sample:\tabular{lcrcr}{
   Exp.Grp \tab Replicate \tab TAXA1 \tab TAXA2 \tab TAXA3 \cr
   X0 \tab 1 \tab 10 \tab 91 \tab 68 \cr
   X0 \tab 2 \tab 11 \tab 86 \tab 70 \cr
   X0 \tab 3 \tab 12 \tab 84 \tab 70 \cr
   X1 \tab 1 \tab 15 \tab 30 \tab 3452 \cr
   X1 \tab 2 \tab 19 \tab 45 \tab 3274 \cr
   X1 \tab 3 \tab 13 \tab 25 \tab 3601 \cr
}
}
\examples{
pairwise_welch(dAtA = CSR_IP2G_data)
pairwise_welch(dAtA = CSR_IP2G_data, var.name = "taxa", p_adj = "BH")
pairwise_welch(dAtA = CSR_IP2G_data, var.name = "taxa", p_adj = "BH", v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
pairwise_welch(dAtA = CSR_TAXA_data)
pairwise_welch(dAtA = CSR_TAXA_data, var.name = "taxa", p_adj = "BH")
pairwise_welch(dAtA = CSR_TAXA_data, var.name = "taxa", p_adj = "BH", v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
}
