% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{summary.CSR_assignments}
\alias{summary.CSR_assignments}
\title{Summary method for CSR_assignments objects}
\usage{
\method{summary}{CSR_assignments}(object)
}
\arguments{
\item{object}{An object of class "CSR_assignments" returned as part of CSR_Results by CSR_assign.}
}
\value{
Invisibly returns a table summarizing the CSR assignments.
}
\description{
Summary method for CSR_assignments objects
}
\examples{
\dontrun{
csr_assignment <- CSR_assign(dAtA = CSR_IP2G_data, var.name = "Trait", p_adj = "BH", Parallel = TRUE)
summary(csr_assignment)
}
}
