% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{summary.CSR_simulation}
\alias{summary.CSR_simulation}
\title{Summary method for CSR_assignments objects}
\usage{
\method{summary}{CSR_simulation}(object)
}
\arguments{
\item{object}{An object of class "CSR_assignments" returned as part of CSR_Results by CSR_assign.}
}
\value{
Invisibly returns a table summarizing the CSR assignments.
}
\description{
Summary method for CSR_assignments objects
}
\examples{
\dontrun{
csr_simulation <- CSR_Simulation(DaTa = CSR_IP2G_data, NSim = 1000, p_adj = "BH", var.name = "Trait", NuLl.test = FALSE, Keep_data = FALSE)
summary.CSR_simulation(csr_simulation)
}
}
