% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{summary.NMA}
\alias{summary.NMA}
\title{Summary method for NMA objects}
\usage{
\method{summary}{NMA}(object)
}
\arguments{
\item{object}{An object of class "NMA".}
}
\value{
Invisibly returns a table summarizing the Null model analysis results.
}
\description{
Summary method for NMA objects
}
\examples{
\dontrun{
  nma_results <- nma_results <- NMA(DaTa = NMA_data, NSim = 1000)
  NMA(nma_results)
}

}
