% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{summary.Pairwise_welch}
\alias{summary.Pairwise_welch}
\title{Summary method for Pairwise_Welch objects}
\usage{
\method{summary}{Pairwise_welch}(object)
}
\arguments{
\item{object}{An object of class "Pariwise_Welch_ANOVA".}
}
\value{
Invisibly returns a table summarizing the Pairwise Welch Anova results.
}
\description{
Summary method for Pairwise_Welch objects
}
\examples{
\dontrun{
pw_results <- pairwise_welch(dAtA = CSR_IP2G_data, var.name = "Traits", p_adj = "BH", 
                      v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
summary(pw_results)
}

}
