% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{summary.Welch_ANOVA}
\alias{summary.Welch_ANOVA}
\title{Summary method for Welch_Anova objects}
\usage{
\method{summary}{Welch_ANOVA}(object)
}
\arguments{
\item{object}{An object of class "Welch_anova".}
}
\value{
Invisibly returns a table summarizing the Welch Anova results.
}
\description{
Summary method for Welch_Anova objects
}
\examples{
\dontrun{
  WA_result <- Welch_ANOVA(dAtA = CSR_IP2G_data, var.name = "Traits", p_adj = "BH", Parallel = TRUE)
  summary.Welch_ANOVA(WA_result)
}

}
