% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{Welch_ANOVA}
\alias{Welch_ANOVA}
\title{Theoretical Microbial Ecology Tools (MicroEcoTools)
Statistical comparison for traits/taxa between experimental groups using Welch-ANOVA or ANOVA.}
\usage{
Welch_ANOVA(dAtA, var.name, p_adj)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}
}
\value{
This function returns a table containing the statistical comparison results.
The input data sample:

| Exp.Grp | Reactor | TAXA | Count |
|-----------|:-----------:|:-----------:|-----------:|
 | I  | 1 | 1 | rk1 |  68  |
 | I  | 2 | 1 | rk1 |  70  |
 | X0 | 1 | 1 | rk1 | 3452 |
 | X0 | 3 | 1 | rk1 | 3274 |
 | X0 | 3 | 1 | rk1 | 3601 |
}
\description{
This function uses oneway test assuming non-equal variance. In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE.
}
\examples{
Welch_ANOVA(dAtA = CSR_data, var.name = "taxa", p_adj = "BH")
}
