% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{CSR_assign}
\alias{CSR_assign}
\title{Theoretical Microbial Ecology Tools (MicroEcoTools)
Assigning CSR categories from Grime's framework to the input variable}
\usage{
CSR_assign(dAtA, var.name, p_adj, p.value.cutoff, Parallel)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}

\item{p.value.cutoff}{A cut-off value for calling the P-value significant can be set using this parameter. The default value is 0.05.}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations.}

\item{v.equal}{Assumption of equal variance can be forced using this parameter. By default it is set to FALSE where it cannot be set to FALSE it will switch to TRUE automatically; a note will be shown in the remarks column in those cases.}
}
\value{
This function returns a table containing the pairwise statistical comparison results. The output table can be fed into the CSR_assign function to assign CSR categories.
The input data sample:

| Exp.Grp | Reactor | TAXA1 | TAXA2 |
|-----------|:-----------:|:-----------:|-----------:|
 | 0  | 1 | 1 | 50 |  68  |
 | 0  | 2 | 1 | 60 |  70  |
 | 0  | 3 | 1 | 10 | 3452 |
 | 1  | 1 | 1 | 56 | 3274 |
 | 1  | 2 | 1 | 89 | 3601 |
 | 1  | 3 | 1 | 56 | 3274 |
}
\description{
This function assigns CSR categories to the input variables using the output table from the ?pairwise_welch function.
To do so, it looks at the pairwise comparisons and assigns C to the ones abundant in no disturbance group, S to the ones abundant in high disturbance group, and R to the ones abundant in intermediate disturbance groups.
It also assigns intermediate categories, CR, CS, and SR to groups that are abundant in no-disturbance and intermediate-disturbance, no-disturbance and high-disturbance, and intermediate-disturbance and high-disturbance, respectively.
Traits that are not differentially abundant in any groups will be assigned to the CSR category.
Please note that if the variability is lacking the group will be categorised as NA.
}
\examples{
CSR_assign(dAtA = CSR_data)
CSR_assign(dAtA = CSR_data, var.name = "taxa", p_adj = "BH")
CSR_assign(dAtA = CSR_data, var.name = "taxa", p_adj = "BH", v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
}
