% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{CSR_assign}
\alias{CSR_assign}
\title{Competitor - Stress tolerant - Ruderal assignment (MicroEcoTools)}
\usage{
CSR_assign(
  dAtA,
  var.name = "Trait",
  p_adj = "BH",
  v.equal = FALSE,
  p.value.cutoff = 0.05,
  Parallel = TRUE,
  Verbose = FALSE
)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function. Defaults to \code{"Trait"}.}

\item{p_adj}{P-value adjustment method. All methods mentioned in \code{\link[stats]{p.adjust}} function can be used ("BH", "BY", "holm", "hommel" or "none"). Defaults to \code{"BH"}.}

\item{v.equal}{Assumption of equal variance can be forced using this parameter. By default it is set to FALSE where it cannot be set to FALSE it will switch to TRUE automatically; a note will be shown in the remarks column in those cases. Defaults to \code{FALSE}.}

\item{p.value.cutoff}{A cut-off value for calling the P-value significant can be set using this parameter.  Defaults to \code{0.05}.}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations. Defaults to \code{TRUE}.}

\item{Verbose}{Run CSR_assign function in verbose mode to explain the assignment. Defaults to \code{FALSE}.}
}
\value{
This function returns a table containing the pairwise statistical comparison results. The output table can be fed into the CSR_assign function to assign CSR categories.
}
\description{
Assigning CSR categories from Grime's framework to the input variable.
This function assigns CSR categories to the input variables using the output table from the \code{\link[MicroEcoTools]{pairwise_welch}} function.
To do so, it looks at the pairwise comparisons and assigns C to the ones abundant in no disturbance group, S to the ones abundant in high disturbance group, and R to the ones abundant in intermediate disturbance groups.
It also assigns intermediate categories, CR, CS, and SR to groups that are abundant in no-disturbance and intermediate-disturbance, no-disturbance and high-disturbance, and intermediate-disturbance and high-disturbance, respectively.
Traits that are not differentially abundant in any groups will be assigned to the CSR category.
Please note that if the variability is lacking the group will be categorised as NA.
}
\details{
The input data sample:\tabular{lcrcr}{
   Exp.Grp \tab Replicate \tab TAXA1 \tab TAXA2 \tab TAXA3 \cr
   X0 \tab 1 \tab 10 \tab 91 \tab 68 \cr
   X0 \tab 2 \tab 11 \tab 86 \tab 70 \cr
   X0 \tab 3 \tab 12 \tab 84 \tab 70 \cr
   X1 \tab 1 \tab 15 \tab 30 \tab 3452 \cr
   X1 \tab 2 \tab 19 \tab 45 \tab 3274 \cr
   X1 \tab 3 \tab 13 \tab 25 \tab 3601 \cr
}
}
\note{
The input data can be relative abundance or count data. The user must ensure the total count across groups is similar.
}
\examples{
# To assign the CSR categories to the traits in the IP2G abundance dataset.
# This example uses abundance data from a perturbation experiment described in Santillan et al. (2019).
# The dataset \code{\link[MicroEcoTools]{CSR_IP2G_data}} is included in the MicroEcoTools package.
# Running the command below runs the welch_anova and pairwise_welch_anova functions.
# This command  automatically assigns C, S, and R traits to traits abundant in undisturbed, highly diturbed and intermediately disturbed experimental groups.
# In addition, it assigns intermediate groups CS, CR, and SR to groups prevalent in two categories, e.g., abundant in undisturbed and intermediately disturbed in compare to highly distrubed --> CR. 
\dontrun{
csr_assignment <- CSR_assign(dAtA = CSR_IP2G_data, var.name = "Trait", p_adj = "BH", Parallel = TRUE)
}

}
