% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{pairwise_welch}
\alias{pairwise_welch}
\title{Pairwise Welch-ANOVA (MicroEcoTools)}
\usage{
pairwise_welch(
  dAtA,
  var.name = "Variable",
  p_adj = "BH",
  v.equal = FALSE,
  p.value.cutoff = 0.05,
  Parallel = TRUE
)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function. Defaults to \code{"Variable"}.}

\item{p_adj}{P-value adjustment method. All methods mentioned in \code{\link[stats]{p.adjust}} function can be used ("BH", "BY", "holm", "hommel" or "none"). Defaults to \code{"BH"}.}

\item{v.equal}{Assumption of equal variance can be forced using this parameter. By default it is set to FALSE where it cannot be set to FALSE it will switch to TRUE automatically; a note will be shown in the remarks column in those cases.Defaults to \code{FALSE}.}

\item{p.value.cutoff}{A cut-off value for calling the P-value significant can be set using this parameter. Defaults to \code{0.05}.}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses 80\% of the available CPU cores to perform the calculations.Defaults to \code{TRUE}}
}
\value{
This function returns a table containing the pairwise statistical comparison results. The output table can be fed into the CSR_assign function to assign CSR categories.
}
\description{
Pairwise comparison between experimental groups using Welch-ANOVA or ANOVA. This function uses oneway test assuming non-equal variance to perform pairwise comparison.
In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE it can be forced to use equal/not equal variance by setting v.equal to TRUE or FALSE.
}
\details{
The input data sample:\tabular{lcrcr}{
   Exp.Grp \tab Replicate \tab TAXA1 \tab TAXA2 \tab TAXA3 \cr
   X0 \tab 1 \tab 10 \tab 91 \tab 68 \cr
   X0 \tab 2 \tab 11 \tab 86 \tab 70 \cr
   X0 \tab 3 \tab 12 \tab 84 \tab 70 \cr
   X1 \tab 1 \tab 15 \tab 30 \tab 3452 \cr
   X1 \tab 2 \tab 19 \tab 45 \tab 3274 \cr
   X1 \tab 3 \tab 13 \tab 25 \tab 3601 \cr
}
}
\note{
The input data can be relative abundance or count data. The user must ensure the total count across groups is similar.
}
\examples{
# Generate a pairwise comparison table for traits across experimental groups.
# This example uses abundance data from a perturbation experiment described in Santillan et al. (2019).
# The dataset \code{\link[MicroEcoTools]{CSR_IP2G_data}} is included in the MicroEcoTools package.
# Running the command below produces a pairwise comparison table for IP2G functions.
# Setting the Parallel to TRUE forces the function to use all available cpu cores to speed up the process.
\dontrun{
pwa_reults <- pairwise_welch(dAtA = CSR_IP2G_data, var.name = "Traits", p_adj = "BH", v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
}

}
