% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{pairwise_welch}
\alias{pairwise_welch}
\title{Theoretical Microbial Ecology Tools (MicroEcoTools)
Pairwise comparison between experimental groups using Welch-ANOVA or ANOVA.}
\usage{
pairwise_welch(dAtA, var.name, p_adj, v.equal, p.value.cutoff, Parallel)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}

\item{v.equal}{Assumption of equal variance can be forced using this parameter. By default it is set to FALSE where it cannot be set to FALSE it will switch to TRUE automatically; a note will be shown in the remarks column in those cases.}

\item{p.value.cutoff}{A cut-off value for calling the P-value significant can be set using this parameter. The default value is 0.05.}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations.}
}
\value{
This function returns a table containing the pairwise statistical comparison results. The output table can be fed into the CSR_assign function to assign CSR categories.
The input data sample:

| Exp.Grp | Reactor | TAXA | Count |
|-----------|:-----------:|:-----------:|-----------:|
 | I  | 1 | 1 | rk1 |  68  |
 | I  | 2 | 1 | rk1 |  70  |
 | X0 | 1 | 1 | rk1 | 3452 |
 | X0 | 3 | 1 | rk1 | 3274 |
 | X0 | 3 | 1 | rk1 | 3601 |
}
\description{
This function uses oneway test assuming non-equal variance to perform pairwise comparison.
In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE it can be forced to use equal/not equal variance by setting v.equal to TRUE or FALSE.
}
\examples{
pairwise_welch(dAtA = CSR_data)
pairwise_welch(dAtA = CSR_data, var.name = "taxa", p_adj = "BH")
pairwise_welch(dAtA = CSR_data, var.name = "taxa", p_adj = "BH", v.equal = FALSE, p.value.cutoff = 0.05, Parallel = TRUE)
}
