% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{Welch_ANOVA}
\alias{Welch_ANOVA}
\title{Welch-ANOVA comparison (MicroEcoTools)}
\usage{
Welch_ANOVA(dAtA, var.name, p_adj, Parallel)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations.}
}
\value{
This function returns a table containing the statistical comparison results.
}
\description{
Statistical comparison for traits/taxa between experimental groups using Welch-ANOVA or ANOVA. This function uses oneway test assuming non-equal variance. In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE.
}
\details{
The input data sample:\tabular{lcrcr}{
   Exp.Grp \tab Reactor \tab TAXA1 \tab TAXA2 \tab TAXA3 \cr
   X0 \tab 1 \tab 10 \tab 91 \tab 68 \cr
   X0 \tab 2 \tab 11 \tab 86 \tab 70 \cr
   X0 \tab 3 \tab 12 \tab 84 \tab 70 \cr
   X1 \tab 1 \tab 15 \tab 30 \tab 3452 \cr
   X1 \tab 2 \tab 19 \tab 45 \tab 3274 \cr
   X1 \tab 3 \tab 13 \tab 25 \tab 3601 \cr
}
}
\examples{
Welch_ANOVA(dAtA = CSR_IP2G_data, var.name = "taxa", p_adj = "BH", Parallel)
Welch_ANOVA(dAtA = CSR_TAXA_data, var.name = "taxa", p_adj = "BH", Parallel)
}
