% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{Welch_ANOVA}
\alias{Welch_ANOVA}
\title{Theoretical Microbial Ecology Tools (MicroEcoTools)
Statistical comparison for traits/taxa between experimental groups using Welch-ANOVA or ANOVA.}
\usage{
Welch_ANOVA(dAtA, var.name, p_adj, Parallel)
}
\arguments{
\item{dAtA}{Dataframe that contains experimental group, replicates, and parameters.}

\item{var.name}{Variable name used for generating output table, for example, taxa, trait, function.}

\item{p_adj}{P-value adjustment method. All methods mentioned in ?p.adjust function can be used ("BH", "BY", "holm", "hommel" or "none").}

\item{Parallel}{Using this parameter you can make use of more CPU cores to decrease run time for the calculation. By default it is set to TRUE and uses half of the available CPU cores to perform the calculations.}
}
\value{
This function returns a table containing the statistical comparison results.
The input data sample:

| Exp.Grp | Reactor | TAXA1 | TAXA2 |
|-----------|:-----------:|:-----------:|-----------:|
 | 0  | 1 | 1 | 50 |  68  |
 | 0  | 2 | 1 | 60 |  70  |
 | 0  | 3 | 1 | 10 | 3452 |
 | 1  | 1 | 1 | 56 | 3274 |
 | 1  | 2 | 1 | 89 | 3601 |
 | 1  | 3 | 1 | 56 | 3274 |
 
 #' @examples
Welch_ANOVA(dAtA = CSR_data, var.name = "taxa", p_adj = "BH", Parallel)
}
\description{
This function uses oneway test assuming non-equal variance. In cases where the assumptions are not valid it chooses equal variance by automatically setting var.equal parameter to TRUE.
}
