% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{Welch_ANOVA}
\alias{Welch_ANOVA}
\title{Welch-ANOVA comparison (MicroEcoTools)}
\usage{
Welch_ANOVA(dAtA, var.name = "Variable", p_adj = "BH", Parallel = TRUE)
}
\arguments{
\item{dAtA}{Data frame containing the experimental group, replicates, and measurement columns.}

\item{var.name}{Character string specifying the variable name to be used in the output table. Defaults to \code{"Variable"}.}

\item{p_adj}{Character string specifying the p-value adjustment method (as per \code{\link[stats]{p.adjust}}). Defaults to \code{"BH"}.}

\item{Parallel}{Logical value. If TRUE, the function uses parallel processing (80% of available CPU cores) to speed up calculations.
Defaults to \code{TRUE}.}
}
\value{
A data frame (table) containing the following columns:
        - The name of each trait/taxa,
        - Its Welch-ANOVA p-value,
        - The test method used,
        - A remarks column indicating if the fallback equal variance test was applied,
        - An adjusted p-value column.
}
\description{
Statistical comparison for traits/taxa between experimental groups using Welch-ANOVA (or ANOVA when variance assumptions fail).
             This function uses a oneway test with the assumption of non-equal variances and automatically falls back to the equal-variance test if needed.
}
\details{
The input data sample is expected to be structured as follows:

| Exp.Grp | Replicate | TAXA1 | TAXA2 | TAXA3 |
|---------|:---------:|------:|------:|------:|
| X0      | 1         | 10    | 91    | 68    |
| X0      | 2         | 11    | 86    | 70    |
| X0      | 3         | 12    | 84    | 70    |
| X1      | 1         | 15    | 30    | 3452  |
| X1      | 2         | 19    | 45    | 3274  |
| X1      | 3         | 13    | 25    | 3601  |
}
\examples{
\dontrun{
  # Using the CSR_IP2G_data dataset included in the package to generate a comparison table.
  wa_results <- Welch_ANOVA(dAtA = CSR_IP2G_data, var.name = "Traits", p_adj = "BH", Parallel = TRUE)
}

}
